Descriptions of informal interfaces (see Pigeons.@informal to see how this page was generated).

log_potential

Description

Pigeons.log_potential — Constant

A log_potential encodes a probability distribution, where only the un-normalized probability density function is known.

To make MyType conforms this informal interface, implement

(log_potential::MyType)(x)

which should return the log of the un-normalized density.

For example, we provide this behaviour for any distribution in Distributions.jl.

log_potentials

Description

Pigeons.log_potentials — Constant

An encoding of a discrete set of probability distributions, where only the un-normalized probability density functions are known. Each distribution is allowed to have a different normalization constant.

For example, we provide this behaviour for any Vector containing log_potential's.

Contract

Pigeons.log_unnormalized_ratio()

pair_swapper

Description

Pigeons.pair_swapper — Constant

Informs swap!() of how to perform a swap between a given pair of chains.

This is done in two steps:

Use swap_stat() to extract sufficient statistics needed to make a swap decision.

Given these statistics for the two chains, swap_decision() then perform the swap.

The rationale for breaking this down into two steps is that in a distributed swap context, swap!() will take care of transmitting the sufficient statistics over the network if necessary.

The function record_swap_stats!() is used to record information about swapping, in particular mean swap acceptance probabilities.

A default implementation of all of pair_swapper's methods is provided, where the pair_swapper is assumed to follow the log_potentials interface.

Contract

Pigeons.swap_stat()

Pigeons.record_swap_stats!()

Pigeons.swap_decision()

Examples of functions providing instances of this interface

Pigeons.TestSwapper()

path

Description

Pigeons.path — Constant

A continuum of log_potential's interpolating between two end-points. More precisely, a mapping from [0, 1] to the space of probability distributions.

The main use of this interface is to pass it to discretize().

Contract

Pigeons.interpolate()

Examples of functions providing instances of this interface

Pigeons.create_path()

recorder

Description

Pigeons.recorder — Constant

Accumulate a specific type of statistic, for example by keeping constant size sufficient statistics (via OnlineStat, which conforms this interface), storing samples to a file, etc. See also recorders.

Contract

Pigeons.record!()

Pigeons.combine()

recorders

Description

Pigeons.recorders — Constant

A NamedTuple containing several recorder's. Each recorder is responsible for a type of statistic to be accumulated.

The keyset of the NamedTuple controls which types of statistics to accumulate (we refer to each element in this keyset as a recorder_key). By default, only those with constant memory requirement are selected, the user can select more expensive ones by enlarging that keyset.

During PT execution, each recorders object keep track of only the statistics for one replica (for thread safety and/or distribution purpose). After a PT round, use reduced_recorders() to do a reduction before accessing statistic values.

Contract

Pigeons.record_if_requested!()

Examples of functions providing instances of this interface

Pigeons.custom_recorders()

Pigeons.default_recorders()

replicas

Description

Pigeons.replicas — Constant

Stores the process' replicas. Since we provide MPI implementations, do not assume that this will contain all the replicas, as others can be located in other processes/machines

Implementations provided

EntangledReplicas: using an MPI-based implementation

Vector{Replica}: for the single process case (above can handle that case, but the array based implementation is non-allocating)

Contract

Pigeons.swap!()

Pigeons.locals()

Pigeons.load()

Pigeons.communicator()

Pigeons.entangler()

Examples of functions providing instances of this interface

Pigeons.create_entangled_replicas()

Pigeons.create_vector_replicas()

state_initializer

Description

Pigeons.state_initializer — Constant

Determine how to initialize the states in the replicas. Implementations include Ref(my_state), to signal all replicas will be initalized to my_state, or a Vector(...) for chain-specific initializations.

Contract

Pigeons.initialization()

swap_graph

Description

Pigeons.swap_graph — Constant

Informs swap!() about which chain will interact with which.

Canonical example is the standard Odd and Even swap, extension point for e.g.

parallel parallel tempering

variational methods with more than 2 legs,

PT algorithms dealing with more than one target simultaneously for the purpose of model selection.

Contract

Pigeons.partner_chain()

Pigeons.reference_chains()

Pigeons.target_chains()

Examples of functions providing instances of this interface

Pigeons.deo()