Descriptions of informal interfaces.

explorer

Description

Pigeons.explorer — Constant

Orchestrate the explore!() phase of Parallel Tempering. This is the part of the algorithm where each replica performs MCMC moves targeting its annealed distribution.

Contract

Pigeons.step!()

Pigeons.adapt_explorer()

Pigeons.explorer_recorder_builders()

log_potential

Description

Pigeons.log_potential — Constant

A log_potential encodes a probability distribution, where only the un-normalized probability density function is known.

To make MyType conform to this informal interface, implement

(log_potential::MyType)(x)

which should return the log of the un-normalized density.

For example, we provide this behaviour for any distribution in Distributions.jl.

Examples of functions providing instances of this interface

Pigeons.create_reference_log_potential()

log_potentials

Description

Pigeons.log_potentials — Constant

An encoding of a discrete set of probability distributions, where only the un-normalized probability density functions are known. Each distribution is allowed to have a different normalization constant.

For example, we provide this behaviour for any Vector containing log_potential's.

Contract

Pigeons.log_unnormalized_ratio()

Pigeons.n_chains()

pair_swapper

Description

Pigeons.pair_swapper — Constant

Informs swap!() of how to perform a swap between a given pair of chains.

This is done in two steps:

Use swap_stat() to extract sufficient statistics needed to make a swap decision.

Given these statistics for the two chains, swap_decision() then perform the swap.

The rationale for breaking this down into two steps is that in a distributed swap context, swap!() will take care of transmitting the sufficient statistics over the network if necessary.

The function record_swap_stats!() is used to record information about swapping, in particular mean swap acceptance probabilities.

A default implementation of all of pair_swapper's methods is provided, where the pair_swapper is assumed to follow the log_potentials interface.

Contract

Pigeons.swap_stat()

Pigeons.record_swap_stats!()

Pigeons.swap_decision()

Examples of functions providing instances of this interface

Pigeons.TestSwapper()

path

Description

Pigeons.path — Constant

A continuum of log_potential's interpolating between two end-points. More precisely, a mapping from [0, 1] to the space of probability distributions.

The main use of this interface is to pass it to discretize().

Contract

Pigeons.interpolate()

Examples of functions providing instances of this interface

Pigeons.InterpolatingPath()

Pigeons.ScaledPrecisionNormalPath()

recorder

Description

Pigeons.recorder — Constant

Accumulate a specific type of statistic, for example by keeping constant size sufficient statistics (via OnlineStat, which conforms this interface), storing samples to a file, etc.

In addition to the contract below, a recorder should support

Base.merge()

Base.empty!()

See also recorders.

Contract

Pigeons.record!()

Examples of functions providing instances of this interface

Pigeons.timing_extrema()

Pigeons.round_trip()

Pigeons.energy_ac1()

Pigeons.log_sum_ratio()

Pigeons._transformed_online()

Pigeons.traces()

Pigeons.allocation_extrema()

Pigeons.reversibility_rate()

Pigeons.explorer_acceptance_pr()

Pigeons.swap_acceptance_pr()

Pigeons.explorer_n_steps()

Pigeons.disk()

Pigeons.index_process()

Pigeons.online()

recorder_builder

Description

Pigeons.recorder_builder — Constant

A function such that calling it returns a fresh recorder.

recorders

Description

Pigeons.recorders — Constant

A NamedTuple containing several recorder's. Each recorder is responsible for a type of statistic to be accumulated (e.g. one for swap accept prs, one for round trip info; some are in-memory, some are on file).

During PT execution, each recorders object keep track of only the statistics for one replica (for thread safety and/or distribution purpose). After a PT round, reduce_recorders!() is used to do a reduction before accessing statistic values.

Contract

Pigeons.record_if_requested!()

Examples of functions providing instances of this interface

Pigeons.create_recorders()

replicas

Description

Pigeons.replicas — Constant

Stores the process' replicas. Since we provide MPI implementations, do not assume that this will contain all the replicas, as others can be located in other processes/machines

Implementations provided

EntangledReplicas: an MPI-based implementation

Vector{Replica}: single-process case (above can handle that case, but the array based implementation is non-allocating)

Contract

Pigeons.swap!()

Pigeons.locals()

Pigeons.load()

Pigeons.communicator()

Pigeons.entangler()

Examples of functions providing instances of this interface

Pigeons.create_entangled_replicas()

Pigeons.create_replicas()

Pigeons.create_vector_replicas()

state

Description

Pigeons.state — Constant

The state held in each Parallel Tempering Replica.

Contract

Pigeons.continuous_variables()

Pigeons.discrete_variables()

Pigeons.variable()

Pigeons.update_state!()

Pigeons.extract_sample()

Pigeons.sample_names()

swap_graph

Description

Pigeons.swap_graph — Constant

Informs swap!() about which chain will interact with which.

These are instantiated by swap_graphs.

Canonical example is the standard Odd and Even swap. Extension point for e.g.

parallel parallel tempering,

variational methods with more than 2 legs,

PT algorithms dealing with more than one target simultaneously for the purpose of model selection.

Contract

Pigeons.partner_chain()

Pigeons.is_reference()

Pigeons.is_target()

swap_graphs

Description

Pigeons.swap_graphs — Constant

Creates one swap_graph for each communication iteration.

Contract

Pigeons.create_swap_graph()

Examples of functions providing instances of this interface

Pigeons.deo()

Pigeons.variational_deo()

target

Description

Pigeons.target — Constant

The probability distribution of interest.

Contract

Pigeons.initialization()

Pigeons.default_explorer()

Pigeons.default_reference()

Pigeons.sample_iid!()

Pigeons.create_path()

Examples of functions providing instances of this interface

Pigeons.toy_mvn_target()

tempering

Description

Pigeons.tempering — Constant

Orchestrate the communicate!() phase of Parallel Tempering.

In addition to the methods in the contract below, we also assume the presence of the following fields:

log_potentials

swap_graphs

communication_barriers

path

Contract

Pigeons.adapt_tempering()

Pigeons.tempering_recorder_builders()

Pigeons.create_pair_swapper()

Pigeons.find_log_potential()

Examples of functions providing instances of this interface

Pigeons.create_tempering()

variational

Description

Pigeons.variational — Constant

A variational family of reference distributions. Implementations should also satisfy the log_potential contract and sample_iid!().

Contract

Pigeons.activate_variational()

Pigeons.update_reference!()

Pigeons.variational_recorder_builders()