Index

Pigeons.log_potential

Pigeons.log_potentials

Pigeons.pair_swapper

Pigeons.path

Pigeons.recorder

Pigeons.recorders

Pigeons.replicas

Pigeons.state_initializer

Pigeons.swap_graph

Pigeons.EntangledReplicas

Pigeons.Entangler

Pigeons.LoadBalance

Pigeons.PermutedDistributedArray

Pigeons.Replica

Pigeons.ScaledPrecisionNormalPath

Pigeons.ScaledPrecisionNormalPath

Pigeons.Schedule

Pigeons.Schedule

Pigeons.Schedule

Pigeons.SwapStat

Pigeons.TestSwapper

Pigeons.TranslatedNormalPath

Pigeons.NRPT

Pigeons.acceptanceprobability

Pigeons.all_reduce_deterministically

Pigeons.analytic_cumulativebarrier

Pigeons.combine

Pigeons.communicationbarrier

Pigeons.communicator

Pigeons.computeetas

Pigeons.create_entangled_replicas

Pigeons.create_path

Pigeons.create_vector_replicas

Pigeons.custom_recorders

Pigeons.default_recorders

Pigeons.deo

Pigeons.deo

Pigeons.deoscan

Pigeons.discretize

Pigeons.entangler

Pigeons.find_global_index

Pigeons.find_local_index

Pigeons.find_process

Pigeons.index_process_plot

Pigeons.informal_doc

Pigeons.initialization

Pigeons.interpolate

Pigeons.load

Pigeons.local_exploration

Pigeons.locals

Pigeons.log_unnormalized_ratio

Pigeons.log_unnormalized_ratio

Pigeons.lognormalizingconstant

Pigeons.my_global_indices

Pigeons.my_load

Pigeons.n_chains_global

Pigeons.one_per_host

Pigeons.partner_chain

Pigeons.permuted_get

Pigeons.permuted_set!

Pigeons.record!

Pigeons.record!

Pigeons.record!

Pigeons.record_if_requested!

Pigeons.record_swap_stats!

Pigeons.record_swap_stats!

Pigeons.recorder_keys

Pigeons.reduce_deterministically

Pigeons.reduced_recorders

Pigeons.reference_chains

Pigeons.restarts

Pigeons.roundtrip

Pigeons.scaled_normal_example

Pigeons.setkernels

Pigeons.single_process_load

Pigeons.slice_accept

Pigeons.slice_double

Pigeons.slice_sample

Pigeons.slice_shrink

Pigeons.split_slice

Pigeons.swap!

Pigeons.swap!

Pigeons.swap!

Pigeons.swap_decision

Pigeons.swap_decision

Pigeons.swap_stat

Pigeons.swap_stat

Pigeons.target_chains

Pigeons.transmit

Pigeons.transmit!

Pigeons.updateschedule

Pigeons.winsorized_mean

Pigeons.winsorized_std

Pigeons.@informal

Types and functions

Pigeons.EntangledReplicas — Type

An implementation of replicas for distributed PT. Contains:

locals

: The subset of replicas hosted in this process

chain_to_replica_global_indices

: A specialized distributed array that maps chain indices to replica indices (global indices). This corresponds to the mapping $\boldsymbol{j}$ in line 2 of Algorithm 5 in Syed et al, 2021.

Pigeons.Entangler — Type

Assume all the MPI processes linked by this communicator will all call the key operations listed below the same number of times in their lifetime, at logically related occasions (e.g. a set number of times per iterations for algorithms running the same number of iterations). We call these 'occasions' a micro-iteration.

This datastructure keeps track internally of appropriate unique tags to coordinate the communication between MPI processes without having to do any explicit synchnonization.

This struct contains:

communicator

: An MPI Comm object (or nothing if a single process is involved).

load

: How a set of tasks or "global indices" are distributed across processes.

current_received_bits

: An internal datastructure used during MPI calls.

n_transmits

: The current micro-iteration.

The key operations supported:

transmit() and transmit!(): encapsulates pairwise communications in which each MPI process is holding a Vector, the elements of which is to be permuted across the processes.

all_reduce_deterministically and reduce_deterministically, to perform MPI collective reduction while maintaining the Parallelism Invariance property.

Pigeons.LoadBalance — Type

Split a list of indices across processes. These indices are denoted $1, 2, .., N$. They are usually some kind of task, for example in the context of parallel tempering, two kinds of tasks arise:

in replicas.state, task $i$ consists in keeping track of the state of replica $i$.

in replicas.chain_to_replica_global_indices, task $i$ consists in storing which replica index corresponds to chain $i$.

One such task index is called a global_index.

LoadBalance splits the global indices among n_processes. LoadBalance is constructed so that the difference in the number of global indices a process is responsible of (its "load") is at most one.

A LoadBalance contains:

my_process_index

: A unique index for this process. We use 1-indexed, i.e. hide MPI's 0-indexed ranks.

n_processes

: Total number of processes involved.

n_global_indices

: The total number of global indices shared between all the processes.

The set {1, 2, .., load()} is called a set of local indices. A local index indexes a slice in {1, 2, ..., n_global_indices}. Collectively over the n_processes, these slices form a partition of the global indices.

Key functions to utilize a LoadBalance struct:

my_global_indices()

find_process()

find_local_index()

my_load()

Pigeons.PermutedDistributedArray — Type

A distributed array making special assumptions on how it will be accessed and written to. The indices of this distributed array correspond to the notion of "global indices" defined in LoadBalance. Several MPI processes cooperate, each processing storing data for a slice of this distributed array.

We make the following assumptions:

Each MPI process will set/get entries the same number of times in their lifetime, at logically related episodes (e.g. a set number of times per iterations for algorithms running the same number of iterations). These episodes are called micro-iteration as in Entangler, which this datastructure is built on.

Moreover, at each time all process perform a get or a set, we assume that each global index is manipulated by exactly one process (i.e. an implicit permutation of the global indices).

We use these assumptions to achieve read/write costs that are near-constant in the number of machines participating.

This struct contains:

local_data

: The slice of the distributed array maintained by this MPI process.

entangler

: An Entangler used to coordinate communication.

The operations supported are:

permuted_get()

permuted_set!()

Pigeons.Replica — Type

One of the $N$ components that forms the state maintained by a PT algorithm. A Replica contains:

state

: Configuration in the state space.

chain

: The index of the distribution currently associated with this replica, modified during swaps.

rng

: Random operations involving this state should use only this random number generator.

recorders

: Records statistics. Each replica carries its own for thread safety/distribution, to be reduced to access.

replica_index

: A global id associated with this replica.

Pigeons.ScaledPrecisionNormalPath — Type

A path of zero-mean normals for testing; contains:

precision0

: Precision parameter of the reference.

precision1

: Precision parameter of the target.

dim

: Dimensionality.

Pigeons.ScaledPrecisionNormalPath — Method

ScaledPrecisionNormalPath(
    dim::Int64
) -> ScaledPrecisionNormalPath

Toy path for testing: see section I.4.1 in Syed et al 2021.

Pigeons.Schedule — Type

A partition of [0, 1] encoded by monotonically increasing grid points starting at zero and ending at one.

Pigeons.Schedule — Method

Schedule(
    n_chains::Int64,
    cumulativebarrier
) -> Schedule{Vector{Float64}}

Create a Schedule with n_chains grid points computed using Algorithm 2 in Syed et al, 2021.

Pigeons.Schedule — Method

Schedule(
    n_chains::Int64
) -> Schedule{StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}

Create a Schedule with n_chains equally spaced grid points.

Pigeons.SwapStat — Type

Default statistics exchanged by a pair of chains in the process of proposing a swap:

log_ratio

uniform

See swap_stat()

Pigeons.TestSwapper — Type

For testing/benchmarking purposes, a simple pair_swapper where all swaps have equal acceptance probability.

Could also be used to warm-start swap connections during exploration phase by setting that constant probability to zero.

Pigeons.TranslatedNormalPath — Type

mean

: Path between a MVN with mean zero at one end point and given mean at the other.

Toy path for testing.

Pigeons.NRPT — Method

NRPT(V_0, V_1, initial_state, ntotal, N)

Non-reversible parallel tempering (NRPT).

Arguments

potential: Function with three arguments (x, η, params) that returns a 'double'. 'x' is the point at which the log-density V0(x; params=params) * η[1] + V1(x) * η[2] is evaluated, where V0 is the negative log density of the reference and V1 is the negative log density of the target.

initial_state: Matrix of initial states for all N+1 chains. Dimensions: (N+1) x (dim_x).

ntotal: Total number of scans/iterations.

N: The total number of chains is N+1.

optimreference: Whether the reference distribution is to be optimized.

maxround: Maximum number of rounds for tuning.

fulltrajectory: Controls whether to keep track of all 'states', 'indices', 'energies', and 'lifts'.

ϕ: (Partially removed. Useful for constructing non-linear paths.)

resolution: Resolution of the output for the estimates of the local communication barrier.

prior_sampler: User may supply an efficient sampler that can obtain samples from the prior / original reference distribution.

optimreference_start: On which tuning round to start optimizing the reference distribution.

full_covariance: Controls whether to use a mean-field approximation for the modified reference (false) or a full covariance matrix (true)

winsorize: Whether or not to use a winsorized/trimmed mean when estimating

the parameters of the variational reference

two_references: Whether to run two PT chains in parallel with two different references: prior and variational reference. Note that with this setting there are 2*(N+1) chains in total.

modref_means_start: Starting values for modref_means

modref_stds_start: Starting values for modref_stds

n_explore: Number of exploration steps to take before considering a communication swap

Pigeons.acceptanceprobability — Method

acceptanceprobability(newenergy, newenergy1, newenergy2)

Compute acceptance probabilities for communication moves. newenergy inputs are lists of numbers.

Arguments

newenergy: -log([πβ0(x^0), πβ1(x^1), ..., πβN(x^N)]) : length N+1

newenergy1: -log([πβ0(x^1), πβ1(x^2), ..., πβ{N-1}(x^N)]) : length N

newenergy2: -log([πβ1(x^0), πβ2(x^1), ..., πβN(x^{N-1})]) : length N

Pigeons.all_reduce_deterministically — Method

Same as reduce_deterministically() except that the result at the root of the tree is then broadcast to all machines so that the output of all_reduce_deterministically() is the root of the reduction tree for all MPI processes involved.

Pigeons.analytic_cumulativebarrier — Method

analytic_cumulativebarrier(
    path::ScaledPrecisionNormalPath
) -> Pigeons.var"#cumulativebarrier#30"{ScaledPrecisionNormalPath, Float64}

Known cumulative barrier used for testing, from Predescu et al., 2003.

Pigeons.combine — Method

combine(recorder1, recorder2) -> Any

Combine the two provided recorder objects.

By default, call Base.merge().

Pigeons.communicationbarrier — Method

communicationbarrier(
    rejection::AbstractVector,
    schedule::AbstractVector
) -> NamedTuple{(:localbarrier, :cumulativebarrier, :globalbarrier), _A} where _A<:Tuple{Pigeons.var"#localbarrier#34", Pigeons.var"#cumulativebarrier#33", Any}

Compute the local communication barrier and cumulative barrier functions from the rejection rates and the current annealing schedule. The estimation of the barriers is based on Fritsch-Carlson monotonic interpolation.

Returns a NamedTuple with fields:

localbarrier

cumulativebarrier

globalbarrier

Pigeons.communicator — Method

communicator(replicas)

Return the replicas's MPI.Comm or nothing if no MPI needed

Pigeons.computeetas — Method

computeetas(ϕ, β) -> Any

Compute the etas matrix given ϕ, which is an Array(K - 1, 2) containing knot parameters, and β, a vector of N+1 schedules. For linear paths, the function returns an (N+1)x2 matrix with entries 1-β in the first column and β in the second column. (This function is useful for those wishing to consider non-linear paths. However, full support is provided only for linear paths at the moment.)

Pigeons.create_entangled_replicas — Function

create_entangled_replicas(
    n_chains::Int64,
    state_initializer,
    rng::SplittableRandoms.SplittableRandom
) -> EntangledReplicas
create_entangled_replicas(
    n_chains::Int64,
    state_initializer,
    rng::SplittableRandoms.SplittableRandom,
    useMPI::Bool
) -> EntangledReplicas
create_entangled_replicas(
    n_chains::Int64,
    state_initializer,
    rng::SplittableRandoms.SplittableRandom,
    useMPI::Bool,
    recorder_keys::Set{Symbol}
) -> EntangledReplicas

Create distributed replicas. The argument useMPI = false is only for debugging purpose. See also state_initializer.

Pigeons.create_path — Function

create_path(ref, target) -> Any
create_path(ref, target, interpolator) -> Any

Given a reference log_potential and a target log_potential, return a path interpolating between the two.

By default, the interpolator is a LinearInterpolator, i.e. standard annealing.

Pigeons.create_vector_replicas — Function

create_vector_replicas(
    n_chains::Int64,
    state_initializer,
    rng::SplittableRandoms.SplittableRandom
) -> Any
create_vector_replicas(
    n_chains::Int64,
    state_initializer,
    rng::SplittableRandoms.SplittableRandom,
    recorder_keys::Set{Symbol}
) -> Any

Create replicas when distributed computing is not needed. See also state_initializer.

Pigeons.custom_recorders — Method

custom_recorders(
    recorder_keys::Set{Symbol}
) -> Union{NamedTuple{(:swap_acceptance_pr,), Tuple{OnlineStatsBase.GroupBy{Tuple{Int64, Int64}, Number, OnlineStatsBase.Mean{Float64, OnlineStatsBase.EqualWeight}}}}, NamedTuple{(:swap_acceptance_pr, :index_process), Tuple{OnlineStatsBase.GroupBy{Tuple{Int64, Int64}, Number, OnlineStatsBase.Mean{Float64, OnlineStatsBase.EqualWeight}}, Dict{Int64, Vector{Int64}}}}}

This returns default_recorders() plus those provided in the recorder_keys.

See also recorder_keys.

Pigeons.default_recorders — Method

default_recorders(

) -> NamedTuple{(:swap_acceptance_pr,), Tuple{OnlineStatsBase.GroupBy{Tuple{Int64, Int64}, Number, OnlineStatsBase.Mean{Float64, OnlineStatsBase.EqualWeight}}}}

Basic, constant-memory recorders.

Pigeons.deo — Method

deo(potential, initial_state, initial_index, initial_lift, schedule, ϕ, 
    nscan, N, resolution, optimreference_round, modref_means, modref_stds, modref_covs, 
    full_covariance, prior_sampler, n_explore)

Deterministic even-odd parallel tempering (DEO/NRPT).

Arguments

potential: Function as in NRPT, but with only two arguments: x and η

initial_state: Starting state, as in NRPT. Input is of size: N+1 [ dim_x ]

initial_index: Starting indices

initial_lift: Starting lift

schedule: Annealing schedule

ϕ: As in NRPT

nscan: Number of scans to use

N: As in NRPT

resolution: As in NRPT

optimreference_round: As in NRPT

modref_means

modref_stds

prior_sample

Pigeons.deo — Method

deo(
    n_chains::Int64,
    current_iteration::Int64
) -> Pigeons.OddEven

Implements the Deterministic Even Odd (DEO) scheme proposed in Okabe, 2001 and analyzed in Syed et al., 2021.

Pigeons.deoscan — Method

deoscan(potential, state, index, lift, etas, n, N, kernels, 
    optimreference_round, modref_means, modref_stds, modref_covs, full_covariance, 
    prior_sampler, n_explore)

Perform one DEO scan (local exploration + communication). Arguments are similar to those for deo(). Note that state is the state from the one previous scan, which is of size N+1[dim_x].

Pigeons.discretize — Method

discretize(path, betas::Schedule) -> Any

Create log_potentials from a path by interpolating the path at each grid point specified in the Schedule.

Pigeons.entangler — Method

entangler(replicas) -> Entangler

Return the replicas's Entangler (possibly a no-communication Entangler if a single process is involved)

Pigeons.find_global_index — Method

find_global_index(
    lb::Pigeons.LoadBalance,
    local_idx::Int64
) -> Int64

Find the global index corresponding to the given local_index.

Pigeons.find_local_index — Method

find_local_index(
    lb::Pigeons.LoadBalance,
    global_idx::Int64
) -> Int64

Find the local index corresponding to the given global_index. Assumes the given global_index is one of this process'.

Pigeons.find_process — Method

find_process(
    lb::Pigeons.LoadBalance,
    global_idx::Int64
) -> Int64

Find the process id (1-indexed) responsible for the given global_idx.

Pigeons.index_process_plot — Method

index_process_plot(recorders) -> Plots.Plot

Given a recorders, create an index process plot.

Pigeons.informal_doc — Method

informal_doc(doc_dir, mod::Module) -> String

Generate informal interface documentation, e.g.:

makedocs(;
    ...
    pages=[
        "Home" => "index.md", 
        "Interfaces" => informal_doc(@__DIR__, MyModuleName),
        ...
    ]
)

Pigeons.initialization — Method

initialization(
    state_initializer,
    rng::SplittableRandoms.SplittableRandom,
    chain::Int64
) -> Any

Determine state_initializer's initialization for the given chain.

Pigeons.interpolate — Method

interpolate(path, beta) -> Pigeons.InterpolatedLogPotential

Returns the log_potential at point beta in the path

Pigeons.load — Method

load(replicas) -> Pigeons.LoadBalance

Return the replicas's LoadBalance (possibly single_process_load)

Pigeons.local_exploration — Method

local_exploration(states, kernels, optimreference_round, modref_means, modref_stds, 
    modref_covs, full_covariance, prior_sampler, n_explore)
    chainacceptance = Vector{Int64}(undef, length(states))

Perform one local exploration move. state is the state from the one previous scan, which is of size N+1[dim_x].

Pigeons.locals — Method

locals(replicas) -> Vector

Return the replicas that are stored in this machine

Pigeons.log_unnormalized_ratio — Method

log_unnormalized_ratio(
    log_potentials::AbstractVector,
    numerator::Int64,
    denominator::Int64,
    state
) -> Any

Assumes the input log_potentials is a vector where each element is a log_potential.

This default implementation is sufficient in most cases, but in less standard scenarios, e.g. where the state space is infinite dimensional, this can be overridden.

Pigeons.log_unnormalized_ratio — Method

log_unnormalized_ratio(
    log_potentials,
    numerator::Int64,
    denominator::Int64,
    state
) -> Any

The argument numerator selects one distribution $\pi_i$ from the collection log_potentials, and similarly denominator selects $\pi_j$. Let $x$ denote the input state. The ratio:

\[f(x) = \frac{\text{d}\pi_i}{\text{d}\pi_j}(x)\]

may only be known up to a normalization constant which can depend on $i$ and $j$ but not $x$, $g(x) = C_{i,j} f(x)$.

This function should return $\log g$ evaluated at state.

Pigeons.lognormalizingconstant — Method

lognormalizingconstant(energies, schedule) -> Any

Compute an estimate of the log normalizing constant given a vector of energies and the corresponding annealing schedule.

Pigeons.my_global_indices — Method

my_global_indices(
    lb::Pigeons.LoadBalance
) -> UnitRange{Int64}

The slice of lb.global_indices this process is reponsible of.

Pigeons.my_load — Method

my_load(lb::Pigeons.LoadBalance) -> Int64

Return the number of indices (task) this process is responsible of.

Pigeons.n_chains_global — Method

n_chains_global(replicas) -> Int64

Given a replicas, return the total number of chains across all processes.

Pigeons.one_per_host — Method

For benchmarking purpose: subset the communicator so that at most one MPI process runs in each machine.

Division is done so that original rank 0 is always included.

Return the new communicator or nothing if this machine is not in the subset.

See also '-s' option in mpi-run

Pigeons.partner_chain — Method

partner_chain(swap_graph, chain::Int64) -> Int64

For a given swap_graph and input chain index, what chain will it interact with at the current iteration? Convention: if a chain is not interacting, return its index.

Pigeons.permuted_get — Method

permuted_get(
    p::PermutedDistributedArray{T},
    indices::AbstractVector{Int64}
) -> Vector

Retreive the values for the given indices, using MPI communication when needed.

We make the following assumptions:

length(indices) == my_load(p.entangler.load)

the indices across all participating processes form a permutation of the global indices.

Pigeons.permuted_set! — Method

permuted_set!(
    p::PermutedDistributedArray{T},
    indices::AbstractArray{T, 1},
    new_values::AbstractArray{T, 1}
)

Set the values for the given indices to the given new_values, using MPI communication when needed.

We make the same assumptions as in permuted_get().

Pigeons.record! — Method

record!(recorder, value) -> Any

Add value to the statistics accumulated by recorder.

Pigeons.record! — Method

record!(recorder::OnlineStatsBase.OnlineStat, value) -> Any

Forwards to OnlineStats' fit!

Pigeons.record! — Method

record!(
    recorder::Dict{K, Array{V, 1}},
    value::Tuple{K, V}
) -> Vector

Given a value, a pair (a, b), and a Dict{K, Vector{V}} backed recorder, append b to the vector corresponding to a, inserting an empty vector into the dictionary first if needed.

Pigeons.record_if_requested! — Method

record_if_requested!(
    recorders::NamedTuple,
    recorder_key,
    value
) -> Any

If the recorders contains the given recorder_key, send the value to the recorder corresponding to the recorder_key. Otherwise, do nothing.

Pigeons.record_swap_stats! — Method

record_swap_stats!(
    pair_swapper,
    recorders,
    chain1::Int64,
    stat1,
    chain2::Int64,
    stat2
)

Given a pair_swapper, a recorders, the provided chain indices, and the sufficient statistics computed by swap_stat(), record statistics.

To avoid accumulating twice the same statistic with (chain1, chain2) and (chain2, chain2), swap!() only calls this for the pair with chain1 < chain2.

By default, the following are accumulated:

the swap acceptance probability.

TODO: stepping stone statistics.

Pigeons.record_swap_stats! — Method

record_swap_stats!(
    swapper::Pigeons.TestSwapper,
    recorder,
    chain1::Int64,
    stat1,
    chain2::Int64,
    stat2
)

See TestSwapper.

Pigeons.recorder_keys — Method

recorder_keys(args::Symbol...) -> Set

Some statistics may induce memory requirements growing in the number of iterations. Use this to select which ones, if any to pass to custom_recorders(). E.g.: recorder_keys() or recorder_keys(:index_process).

Choices include (each specifying if it is included in default_recorders()):

:swap_acceptance_pr: maintain swap acceptance probabilities, a GroupBy(Tuple{Int, Int}, Mean()) object (included by default);

:index_process: keep, for each replica, the list of chains visited (not included by default), a Dict{Int, Vector{Int}} object.

Pigeons.reduce_deterministically — Method

reduce_deterministically(
    operation,
    source_data::AbstractArray{T, 1},
    e::Entangler
) -> Any

Perform a binary reduction of the source_data, using MPI when needed.

Consider the binary tree with leaves given by the global indices specified in e.load and stored in the different MPI processes' input source_data vectors. At each node of the tree, a reduction is performed using operation, i.e. by calling operation(left_child, right_child). When, and only when a branch of the tree crosses from one MPI process to another one, MPI communication is used to transmit the intermediate reduction.

At the end, for process 1, reduce_deterministically() will return the root of the binary tree, and for the other processes, reduce_deterministically() will return nothing.

Note that even when the operation is only approximately associative (typical situation for floating point reductions), the output of this function is invariant to the number of MPI processes involved (hence the terminology 'deterministically'). This contrasts to direct use of MPI collective communications where the leaves are MPI processes and hence will give slightly different outputs given different numbers of MPI processes. In the context of randomized algorithms, these minor differences are then amplified.

In contrast to transmit!(), we do not assume isbitstype(T) == true and use serialization when messages are transmitted over MPI.

Pigeons.reduced_recorders — Method

reduced_recorders(replicas) -> Any

Perform a reduction across all the replicas' individual recorders, using combine() on each individual recorder held. Returns a recorders with all the information merged.

Since this uses all_reduce_deterministically, the output is identical, no matter how many MPI processes are used, even when the reduction involves only approximately associative combine() operations (e.g. most floating point ones).

Pigeons.reference_chains — Method

reference_chains(swap_graph) -> Set{Int64}

Given a swap_graph, return the set of chain(s) targeting a reference distribution. These are typically tractable in the sense that we can sample i.i.d. from them.

Pigeons.restarts — Method

restarts(indices_matrix; cumulative) -> Any

Compute the number of restarts for a given index process trajectory. Otherwise, it is the same as roundtrip().

Pigeons.roundtrip — Method

roundtrip(indices_matrix; cumulative) -> Any

Compute the number of round trips for a given index process trajectory. indices_matrix is a matrix containing information about the index process. cumulative indicates whether we should store the output as a vector containing information about the number of total round trips up to sample n. If false, the output is a scalar.

Pigeons.scaled_normal_example — Method

scaled_normal_example(n_chains, dim) -> Any

Toy path for testing: see section I.4.1 in Syed et al 2021.

Pigeons.setkernels — Method

setkernels(potential, etas)

Set the local exploration kernels given the potential and the annealing parameters, etas.

Pigeons.single_process_load — Method

single_process_load(n_global_indices) -> Pigeons.LoadBalance

A load balance with only one process.

Pigeons.slice_accept — Method

slice_accept(h::SS, g, x_0, x_1, z, L, R, c)

Test whether to accept the current slice.

Pigeons.slice_double — Method

slice_double(h, g, x_0, z, c)

Double the current slice.

Pigeons.slice_sample — Method

slice_sample(h::SS, x_0::Vector{Float64}, nsamps::Int)

Slice sample nsamps given a starting vector x_0 and the struct h that contains information about the log-density.

Pigeons.slice_shrink — Method

slice_shrink(h, g, x_0, z, L, R, c)

Shrink the current slice.

Pigeons.split_slice — Method

split_slice(slice::UnitRange, rng) -> Any

From one splittable random object, one can conceptualize an infinite list of splittable random objects. Return a slice from this infinite list.

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

For each pair of chains encoded in swap_graph, use pair_swapper to decide if the pair will swap or not, and write the changes in-place into replicas (i.e. exchanging the Replica's chain fields for those that swapped.)

Pigeons.swap! — Method

swap!(pair_swapper, replicas::EntangledReplicas, swap_graph)

Entangled MPI swap! implementation.

This implementation is designed to support distributed PT with the following guarantees

The running time is independent of the size of the state space ('swapping annealing parameters rather than states')

The output is identical no matter how many MPI processes are used. In particular, this means that we can check correctness by comparing to the serial, single-process version.

Scalability to 1000s of processes communicating over MPI (see details below).

The same function can be used when a single process is used and MPI is not available.

Flexibility to extend PT to e.g. networks of targets and general paths.

Running time analysis:

Let $N$ denote the number of chains, $P$, the number of processes, and $K = \text{ceil}(N/P)$, the maximum number of chains held by one process. Assuming the running time is dominated by communication latency and a constant time for the latency of each peer-to-peer communication, the theoretical running time is $O(K)$. In practice, latency will grow as a function of $P$, but empirically, this growth appears to be slow enough that for say $P = N =$ a few 1000s, swapping will not be the computational bottleneck.

Pigeons.swap! — Method

swap!(pair_swapper, replicas::Array{R, 1}, swap_graph)

Single process, non-allocating swap! implementation.

Pigeons.swap_decision — Method

swap_decision(
    pair_swapper,
    chain1::Int64,
    stat1,
    chain2::Int64,
    stat2
) -> Bool

Given a pair_swapper, a recorders, the provided chain indices, and the sufficient statistics computed by swap_stat(), make a swap decision.

By default, this is done as follows:

compute the standard swap acceptance probability min(1, exp(stat1.log_ratio + stat2.log_ratio))

make sure the two chains share the same uniform by picking the uniform from the chain with the smallest chain index

swap if the shared uniform is smaller than the swap acceptance probability.

Pigeons.swap_decision — Method

swap_decision(
    swapper::Pigeons.TestSwapper,
    chain1::Int64,
    stat1::Float64,
    chain2::Int64,
    stat2::Float64
) -> Bool

See TestSwapper.

Pigeons.swap_stat — Method

swap_stat(
    pair_swapper,
    replica::Replica,
    partner_chain::Int64
) -> Float64

By default, two sufficient statistics are computed and stored in the SwapStat struct:

The result of calling log_unnormalized_ratio() on pair_swapper

A uniform number to coordinate the swap decision.

This can be extended by dispatching on other pair_swapper types, with the constraint that the returned sufficient statistics should satisfy isbitstype().

Pigeons.swap_stat — Method

swap_stat(
    swapper::Pigeons.TestSwapper,
    replica::Replica,
    partner_chain::Int64
) -> Float64

See TestSwapper.

Pigeons.target_chains — Method

target_chains(swap_graph) -> Set{Int64}

Given a swap_graph, return the set of chain(s) targeting a distribution of interest.

Pigeons.transmit! — Method

transmit!(
    e::Entangler,
    source_data::AbstractArray{T, 1},
    to_global_indices::AbstractVector{Int64},
    write_received_data_here::Array{T, 1}
)

Use MPI point-to-point communication to permute the contents of source_data across MPI processes, writing the permuted data into write_received_data_here. The permutation is specified by the load balance in the input argument e as well as the argument to_global_indices.

More precisely, assume the Vectors source_data, to_global_indices, and write_received_data_here are all of the length specified in my_load(e.load).

For each i, source_data[i] is sent to MPI process p = find_process(e.load, g), where g = to_global_indices[i] and written into this p 's write_received_data_here[j], where j = find_local_index(e.load, g)

See Entangler's comments regarding the requirement that all machines call transmit() the same number of times and at logically related intervals.

Additionally, at each micro-iteration, we assume that {to_global_indices_p : p ranges over the difference processes} forms a partition of {1, ..., e.load.n_global_indices} If ran in single-process mode, this 'partition property' is checked; if ran in multi-process, opportunistic checks will be made, namely when several entries in to_global_indices lie in the same process, but systematic checks are not made for performance reasons.

We also assume isbitstype(T) == true.

Pigeons.transmit — Method

transmit(
    e::Entangler,
    source_data::AbstractArray{T, 1},
    to_global_indices::AbstractVector{Int64}
) -> Vector

The same as transmit!() but instead of writing the result to an input argument, provide the result as a returned Vector.

Pigeons.updateschedule — Method

updateschedule(
    cumulativebarrier,
    N::Int64
) -> Vector{Float64}

Update the annealing schedule. Given the cumulative communication barrier function in cumulativebarrier, find the optimal schedule of size N+1.

Pigeons.winsorized_mean — Method

winsorized_mean(x; α) -> Any

Compute the winsorized mean from an input x, which is assumed to be a vector of vectors. α denotes the percentage of observations to winsorize at the bottom and the top so that we use 1 - 2α observations and winsorize the rest.

Pigeons.winsorized_std — Method

winsorized_std(x; α) -> Any

Compute the winsorized standard deviation. The parameters are the same as those for winsorized_mean().

Pigeons.@informal — Macro

@informal name begin ... end

Document an informal interface with provided name, and functions specified in a begin .. end block.

@informal will spit back the contents of the begin .. end block so this macro can be essentially ignored at first read.

When building documentation, this allows us to use the function informal_doc() to automatically document the informal interface.