Index

Pigeons.explorer

Pigeons.log_potential

Pigeons.log_potentials

Pigeons.pair_swapper

Pigeons.path

Pigeons.recorder

Pigeons.recorder_builder

Pigeons.recorders

Pigeons.replicas

Pigeons.state_initializer

Pigeons.swap_graph

Pigeons.swap_graphs

Pigeons.target

Pigeons.tempering

Pigeons.ChildProcess

Pigeons.EntangledReplicas

Pigeons.Entangler

Pigeons.FromCheckpoint

Pigeons.Immutable

Pigeons.Indexer

Pigeons.Indexer

Pigeons.Inputs

Pigeons.InterpolatedLogPotential

Pigeons.InterpolatingPath

Pigeons.Iterators

Pigeons.LoadBalance

Pigeons.MPI

Pigeons.MPISettings

Pigeons.NonReversiblePT

Pigeons.NonReversiblePT

Pigeons.PT

Pigeons.PT

Pigeons.PT

Pigeons.PermutedDistributedArray

Pigeons.Replica

Pigeons.Reproducibility

Pigeons.Result

Pigeons.ScaledPrecisionNormalPath

Pigeons.ScaledPrecisionNormalPath

Pigeons.Schedule

Pigeons.Shared

Pigeons.Shared

Pigeons.SliceSampler

Pigeons.Submission

Pigeons.SwapStat

Pigeons.TestSwapper

Pigeons.ThisProcess

Pigeons.ToyExplorer

Pigeons.TuringLogPotential

Pigeons.NRPT

Pigeons.acceptanceprobability

Pigeons.adapt

Pigeons.adapt_explorer

Pigeons.adapt_tempering

Pigeons.adapted_schedule

Pigeons.all_reduce_deterministically

Pigeons.analytic_cumulativebarrier

Pigeons.check_against_serial

Pigeons.checksum

Pigeons.communicate!

Pigeons.communicationbarrier

Pigeons.communicator

Pigeons.computeetas

Pigeons.create_entangled_replicas

Pigeons.create_explorer

Pigeons.create_explorer

Pigeons.create_explorer

Pigeons.create_pair_swapper

Pigeons.create_path

Pigeons.create_path

Pigeons.create_recorders

Pigeons.create_recorders

Pigeons.create_reference_log_potential

Pigeons.create_replicas

Pigeons.create_state_initializer

Pigeons.create_swap_graph

Pigeons.create_tempering

Pigeons.create_vector_replicas

Pigeons.deo

Pigeons.deo

Pigeons.deoscan

Pigeons.deserialize_immutables

Pigeons.discretize

Pigeons.entangler

Pigeons.equally_spaced_schedule

Pigeons.explore!

Pigeons.explore!

Pigeons.explorer_recorder_builders

Pigeons.find_global_index

Pigeons.find_local_index

Pigeons.find_log_potential

Pigeons.find_process

Pigeons.flush_immutables!

Pigeons.index_process

Pigeons.index_process_plot

Pigeons.informal_doc

Pigeons.initialization

Pigeons.interpolate

Pigeons.is_finished

Pigeons.latest_checkpoint_folder

Pigeons.load

Pigeons.load

Pigeons.local_exploration

Pigeons.locals

Pigeons.log_unnormalized_ratio

Pigeons.log_unnormalized_ratio

Pigeons.lognormalizingconstant

Pigeons.mpi_active

Pigeons.my_global_indices

Pigeons.my_load

Pigeons.n_chains

Pigeons.next_exec_folder

Pigeons.one_per_host

Pigeons.only_one_process

Pigeons.partner_chain

Pigeons.permuted_get

Pigeons.permuted_set!

Pigeons.pigeons

Pigeons.pigeons

Pigeons.pigeons

Pigeons.pigeons

Pigeons.pigeons

Pigeons.providers

Pigeons.record!

Pigeons.record!

Pigeons.record!

Pigeons.record_if_requested!

Pigeons.record_swap_stats!

Pigeons.record_swap_stats!

Pigeons.reduce_deterministically

Pigeons.reduce_recorders!

Pigeons.reference_chains

Pigeons.report

Pigeons.restarts

Pigeons.roundtrip

Pigeons.run_checks

Pigeons.run_one_round!

Pigeons.sample_iid!

Pigeons.serialize_immutables

Pigeons.setkernels

Pigeons.setup_mpi

Pigeons.setup_mpi

Pigeons.single_process_load

Pigeons.slice_accept

Pigeons.slice_double

Pigeons.slice_sample!

Pigeons.slice_shrink

Pigeons.sort_includes!

Pigeons.split_slice

Pigeons.step!

Pigeons.swap!

Pigeons.swap!

Pigeons.swap!

Pigeons.swap_acceptance_pr

Pigeons.swap_decision

Pigeons.swap_decision

Pigeons.swap_stat

Pigeons.swap_stat

Pigeons.target_chains

Pigeons.tempering_recorder_builders

Pigeons.toy_mvn_target

Pigeons.transmit

Pigeons.transmit!

Pigeons.updateschedule

Pigeons.winsorized_mean

Pigeons.winsorized_std

Pigeons.write_checkpoint

Pigeons.@abstract

Pigeons.@informal

Pigeons.@weighted

Types and functions

Pigeons.ChildProcess — Type

Flag to run to a new julia process. Useful e.g. to dynamically control the number of threads to use. Fields:

n_threads: The number of threads to provide in the child julia process, the same as the current process by default.

extra_julia_modules: Extra Julia Modules needed by the child process.

n_local_mpi_processes: If greater than one, run the code locally over MPI using that many MPI processes. In most cases, this is useful only for debugging purpose, as multi-threading should typically perform better. This could also potentially be useful if using a third-party target distribution which somehow does not support multi-threading.

wait: If wait is false, the process runs asynchronously. When wait is false, the process' I/O streams are directed to devnull.

Pigeons.EntangledReplicas — Type

An implementation of replicas for distributed PT. Contains:

locals: The subset of replicas hosted in this process

chain_to_replica_global_indices: A specialized distributed array that maps chain indices to replica indices (global indices). This corresponds to the mapping $\boldsymbol{j}$ in line 2 of Algorithm 5 in Syed et al, 2021.

Pigeons.Entangler — Type

Assume all the MPI processes linked by this communicator will all call the key operations listed below the same number of times in their lifetime, at logically related occasions (e.g. a set number of times per iteration for algorithms running the same number of iterations). We call these 'occasions' a micro-iteration.

This datastructure keeps track internally of appropriate unique tags to coordinate the communication between MPI processes without having to do any explicit synchronization.

This struct contains:

communicator: An MPI Comm object (or nothing if a single process is involved).

load: How a set of tasks or "global indices" are distributed across processes.

current_received_bits: An internal datastructure used during MPI calls.

n_transmits: The current micro-iteration.

The key operations supported:

transmit() and transmit!(): encapsulates pairwise communications in which each MPI process is holding a Vector, the elements of which are to be permuted across the processes.

all_reduce_deterministically and reduce_deterministically, to perform MPI collective reduction while maintaining the Parallelism Invariance property.

Pigeons.FromCheckpoint — Type

Flag create_replicas (and related functions) that replicas should be loaded from a checkpoint. Fields:

checkpoint_folder

Pigeons.Immutable — Method

Immutable(data)

Consider a situation where a distributed system serializes its state, and part of the state contains large immutable data. When the distributed processes each independently call Serialization.serialize(), naively the processes would each write identical copies of the large immutable data, which is space-inefficient.

Immutable resolves this space-inefficiency. For most users, all they need to know is to enclose large data inside the struct Immutable.

Details of how serialization/deserialization is performed:

Enclose large immutable data inside a Immutable. Assume the type of data has well defined hash and ==. Internally, we maintain an internal, global Dict indexed by hash(data) storing the data. This global Dict is called immutables.

Use flush_immutable() to clear the global immutable state

Use Serialization.serialize as usual. Internally, we dispatch serialization of Immutable is modified to skip the field containing the data.

Make one of the processes call serialize_immutables(). This serializes the immutables Dict.

Then for de-serialization, each process should call deserialize_immutables(). This restores immutable.

Finally, call Serialization.deserialize() as usual. When an Immutable instances is being deserialize, we dispatch deserialization so that the data is retreived from immutable.

Pigeons.Indexer — Type

A bijection between integers and some type T. T is assumed to have consistent hash and ==. The two sides of the bijection can be obtained with the fields:

i2t: A Vector mapping integers to objects t of type T.

t2i: A Dict mapping objects t of type T to integers.

Pigeons.Indexer — Method

Indexer(i2t)

Pigeons.Inputs — Type

A Base.@kwdef struct used to create Parallel Tempering algorithms.

Fields (see source file for default values):

target: The target distribution.

seed: The master random seed.

n_rounds: The number of rounds to run.

n_chains: The number of chains to use.

checkpoint: Whether a checkpoint should be written to disk at the end of each round.

recorder_builders: An Vector with elements of type recorder_builder.

checked_round: The round index where run_checks() will be performed. Set to 0 to skip these checks.

Pigeons.InterpolatedLogPotential — Type

A log_potential obtained by evaluation of a path at a point beta.

Pigeons.InterpolatingPath — Method

InterpolatingPath(ref, target)

Given a reference log_potential and a target log_potential, return a path interpolating between them.

By default, the interpolator is a LinearInterpolator, i.e. standard annealing.

Pigeons.Iterators — Type

Iterators used in Parallel Tempering. Stored in a struct so that recorder's can access it when outputting sample statistics.

Fields:

round: Index of the Parallel Tempering adaptation round, as defined in Algorithm 4 of Syed et al., 2021. Set to zero when when pigeons() not yet started.

scan: Number of (exploration, communication) pairs performed so far, corresponds to $n$ in Algorithm 1 of Syed et al., 2021. Round $i$ typically performs $2^i$ scans. Set to zero when runoneround!() is not yet started.

Pigeons.LoadBalance — Type

Split a list of indices across processes. These indices are denoted $1, 2, .., N$. They are usually some kind of task, for example in the context of parallel tempering, two kinds of tasks arise:

in replicas.state, task $i$ consists in keeping track of the state of replica $i$.

in replicas.chain_to_replica_global_indices, task $i$ consists in storing which replica index corresponds to chain $i$.

One such task index is called a global_index.

LoadBalance splits the global indices among n_processes. LoadBalance is constructed so that the difference in the number of global indices a process is responsible of (its "load") is at most one.

A LoadBalance contains:

my_process_index: A unique index for this process. We use 1-indexed, i.e. hide MPI's 0-indexed ranks.

n_processes: Total number of processes involved.

n_global_indices: The total number of global indices shared between all the processes.

The set {1, 2, .., load()} is called a set of local indices. A local index indexes a slice in {1, 2, ..., n_global_indices}. Collectively over the n_processes, these slices form a partition of the global indices.

Key functions to utilize a LoadBalance struct:

my_global_indices()

find_process()

find_local_index()

my_load()

Pigeons.MPI — Type

Flag to run on MPI. Before using, you have to call once setup_mpi.

Fields:

n_threads: The number of threads per MPI process, 1 by default.

walltime: The walltime limit, 00:30:00 by default (i.e., 30 minutes).

n_mpi_processes: The number of MPI processes, 2 by default.

memory: The memory allocated to each MPI process, 8gb by default.

extra_julia_modules: Extra Julia Modules needed by the child process.

Pigeons.MPISettings — Type

Global settings needed for MPI job submission:

allocation_code: E.g., for -A in PBS submission scripts.

environment_modules: Run module avail in the terminal to see what is available on your HPC.

Pigeons.NonReversiblePT — Type

Variables needed for the non-reversible Parallel Tempering described in Syed et al., 2021:

path: The path.

schedule: The Schedule.

log_potentials: The log_potentials.

swap_graphs: The swap_graphs.

Pigeons.NonReversiblePT — Method

NonReversiblePT(inputs)

The adaptive non-reversible Parallel Tempering described in Syed et al., 2021.

Pigeons.PT — Type

Storage involved in PT algorithms:

inputs: The user-provided Inputs that determine the execution of a PT algorithm.

replicas: The replicas held by this machine.

shared: Information shared across all machines, updated between rounds.

exec_folder: Either a path to a folder shared by all processes, which is used to save information to disk (checkpoints, samples etc); or nothing if a completely in-memory algorithm is used.

reduced_recorders: recorders from the last round, or empty recorders.

Pigeons.PT — Method

PT(source_exec_folder; round, fresh_exec_folder)

Create a PT struct from a saved checkpoint. The path [checkpoint_folder] should point to a folder with the name checkpoint found under results/all/[exec_folder]/round=x.

The checkpoint carries all the information stored in a PT struct. It is possible for an MPI-based execution to load a checkpoint written by a single-process execution and vice versa.

A new unique folder will be created with symlinks to the source one, so that e.g. running more rounds of PT will results in a new space-efficient checkpoint containing all the information for the new run.

Pigeons.PT — Method

PT(inputs; exec_folder)

Pigeons.PermutedDistributedArray — Type

A distributed array making special assumptions on how it will be accessed and written to. The indices of this distributed array correspond to the notion of "global indices" defined in LoadBalance. Several MPI processes cooperate, each processing storing data for a slice of this distributed array.

We make the following assumptions:

Each MPI process will set/get entries the same number of times in their lifetime, at logically related episodes (e.g. a set number of times per iteration for algorithms running the same number of iterations). These episodes are called micro-iterations as in Entangler, which this datastructure is built on.

Moreover, at each time all processes perform a get or a set, we assume that each global index is manipulated by exactly one process (i.e. an implicit permutation of the global indices).

We use these assumptions to achieve read/write costs that are near-constant in the number of machines participating.

This struct contains:

local_data: The slice of the distributed array maintained by this MPI process.

entangler: An Entangler used to coordinate communication.

The operations supported are:

permuted_get()

permuted_set!()

Pigeons.Replica — Type

One of the $N$ components that forms the state maintained by a PT algorithm. A Replica contains:

state: Configuration in the state space.

chain: The index of the distribution currently associated with this replica, modified during swaps.

rng: Random operations involving this state should use only this random number generator.

recorders: Records statistics. Each replica carries its own for thread safety/distribution; then they are reduced at end of each round.

replica_index: A global id associated with this replica.

Pigeons.Reproducibility — Type

Used to check reproducibility of jobs. Less emphasis on speed, more on getting diagnostic when reproducibility is violated.

Pigeons.Result — Type

A link to an execution folder able to deserialize type T via a string constructor.

Pigeons.ScaledPrecisionNormalPath — Type

A path of zero-mean normals for testing; contains:

precision0: Precision parameter of the reference.

precision1: Precision parameter of the target.

dim: Dimensionality.

Pigeons.ScaledPrecisionNormalPath — Method

ScaledPrecisionNormalPath(dim)

Toy Multivariate Normal (MVN) path of distributions for testing: see section I.4.1 in Syed et al 2021.

Pigeons.Schedule — Type

A partition of [0, 1] encoded by monotonically increasing grid points starting at zero and ending at one.

Pigeons.Shared — Type

Information shared by all processes involved in a round of distributed parallel tempering. This is updated between rounds but only read during a round.

Fields:

iterators: See Iterators.

tempering: See tempering.

explorer: See explorer.

Only one instance maintained per process.

Pigeons.Shared — Method

Shared(inputs)

Create a Shared struct based on an Inputs.

Uses create_tempering() and create_explorer().

Pigeons.SliceSampler — Type

Slice sampler based on Neal, 2003.

Pigeons.Submission — Type

Specifies where to submit a task.

Pigeons.SwapStat — Type

Default statistics exchanged by a pair of chains in the process of proposing a swap:

log_ratio

uniform

See swap_stat()

Pigeons.TestSwapper — Type

For testing/benchmarking purposes, a simple pair_swapper where all swaps have equal acceptance probability.

Could also be used to warm-start swap connections during exploration phase by setting that constant probability to zero.

Pigeons.ThisProcess — Type

Flag to ask to run a function within the current process.

Pigeons.ToyExplorer — Type

Toy explorer for toy paths where each log_potential supports i.i.d. sampling via rand(rng, log_potential).

Pigeons.TuringLogPotential — Method

TuringLogPotential(model)

Pigeons.NRPT — Method

NRPT(V_0, V_1, initial_state, ntotal, N)

Non-reversible parallel tempering (NRPT).

Arguments

potential: Function with three arguments (x, η, params) that returns a 'double'. 'x' is the point at which the log-density V0(x; params=params) * η[1] + V1(x) * η[2] is evaluated, where V0 is the negative log density of the reference and V1 is the negative log density of the target.

initial_state: Matrix of initial states for all N+1 chains. Dimensions: (N+1) x (dim_x).

ntotal: Total number of scans/iterations.

N: The total number of chains is N+1.

optimreference: Whether the reference distribution is to be optimized.

maxround: Maximum number of rounds for tuning.

fulltrajectory: Controls whether to keep track of all 'states', 'indices', 'energies', and 'lifts'.

ϕ: (Partially removed. Useful for constructing non-linear paths.)

resolution: Resolution of the output for the estimates of the local communication barrier.

prior_sampler: User may supply an efficient sampler that can obtain samples from the prior / original reference distribution.

optimreference_start: On which tuning round to start optimizing the reference distribution.

full_covariance: Controls whether to use a mean-field approximation for the modified reference (false) or a full covariance matrix (true)

winsorize: Whether or not to use a winsorized/trimmed mean when estimating

the parameters of the variational reference

two_references: Whether to run two PT chains in parallel with two different references: prior and variational reference. Note that with this setting there are 2*(N+1) chains in total.

modref_means_start: Starting values for modref_means

modref_stds_start: Starting values for modref_stds

n_explore: Number of exploration steps to take before considering a communication swap

Pigeons.acceptanceprobability — Method

acceptanceprobability(newenergy, newenergy1, newenergy2)

Compute acceptance probabilities for communication moves. newenergy inputs are lists of numbers.

Arguments

newenergy: -log([πβ0(x^0), πβ1(x^1), ..., πβN(x^N)]) : length N+1

newenergy1: -log([πβ0(x^1), πβ1(x^2), ..., πβ{N-1}(x^N)]) : length N

newenergy2: -log([πβ1(x^0), πβ2(x^1), ..., πβN(x^{N-1})]) : length N

Pigeons.adapt — Method

adapt(pt, reduced_recorders)

Call adapt_tempering() followed by adapt_explorer.

Pigeons.adapt_explorer — Method

adapt_explorer(explorer, reduced_recorders, shared)

Called between successive rounds (run_one_round!).

Given an explorer, reduced recorders and Shared return an updated explorer.

Pigeons.adapt_tempering — Method

adapt_tempering(tempering, reduced_recorders)

Called between successive rounds (run_one_round!).

Given a tempering and reduced recorders return an updated tempering.

Pigeons.adapted_schedule — Method

adapted_schedule(n_chains, cumulativebarrier)

Create a Schedule with n_chains grid points computed using Algorithm 2 in Syed et al, 2021.

Pigeons.all_reduce_deterministically — Method

all_reduce_deterministically(operation, source_data, e)

Same as reduce_deterministically() except that the result at the root of the tree is then broadcast to all machines so that the output of all_reduce_deterministically() is the root of the reduction tree for all MPI processes involved.

Pigeons.analytic_cumulativebarrier — Method

analytic_cumulativebarrier(path)

Known cumulative barrier used for testing, from Predescu et al., 2003.

Pigeons.check_against_serial — Method

Run a separate, fully serial version of the PT algorithm, and compare the checkpoint files to ensure the two produce exactly the same output.

Pigeons.checksum — Function

checksum(filename)
checksum(filename, blocksize)

Given a filename path, compute a crc32c() checksum in constant memory.

Pigeons.communicate! — Method

communicate!(pt)

Use create_pair_swapper() and create_swap_graph to construct the inputs needed for swap!.

Pigeons.communicationbarrier — Method

communicationbarrier(rejection, schedule)

Compute the local communication barrier and cumulative barrier functions from the rejection rates and the current annealing schedule. The estimation of the barriers is based on Fritsch-Carlson monotonic interpolation.

Returns a NamedTuple with fields:

localbarrier

cumulativebarrier

globalbarrier

Pigeons.communicator — Method

communicator(replicas)

Return the replicas's MPI.Comm or nothing if no MPI needed

Pigeons.computeetas — Method

computeetas(ϕ, β)

Compute the etas matrix given ϕ, which is an Array(K - 1, 2) containing knot parameters, and β, a vector of N+1 schedules. For linear paths, the function returns an (N+1)x2 matrix with entries 1-β in the first column and β in the second column. (This function is useful for those wishing to consider non-linear paths. However, full support is provided only for linear paths at the moment.)

Pigeons.create_entangled_replicas — Method

create_entangled_replicas(inputs, shared, source)

Create distributed replicas.

See create_replicas.

Pigeons.create_explorer — Method

create_explorer(target, inputs)

Pigeons.create_explorer — Method

create_explorer(target, inputs)

Create an explorer for the given target.

Pigeons.create_explorer — Method

create_explorer(inputs)

Given an Inputs object, dispatch on create_explorer(inputs.target, inputs) to construct the explorer associated with the input target distribution.

Pigeons.create_pair_swapper — Method

create_pair_swapper(tempering, target)

Given a tempering and a Shared struct, create a pair_swapper.

If ommitted, by default will return the standard Metropolis-Hastings accept-reject.

Pigeons.create_path — Method

create_path(target, inputs)

Create a path, by default linking the given target to the refence provided by create_reference_log_potential().

For this default to work, the target should conform both target and log_potential.

Pigeons.create_path — Method

create_path(target, inputs)

In this case, the target is already a path, so return it.

Pigeons.create_recorders — Method

create_recorders(recorder_builders)

Create a recorders from an iterable with element type recorder_builder.

Pigeons.create_recorders — Method

create_recorders(inputs, shared)

Create a recorders from an Inputs and Shared.

Pigeons.create_reference_log_potential — Method

create_reference_log_potential(target, inputs)

Create a default reference distribution, by returning a log_potential. The returned object will also get passed to sample_iid!() at the "hot chains" of the Parallel Tempering algorithm.

Pigeons.create_replicas — Method

create_replicas(inputs, shared, source)

Create replicas, detecting automatically if MPI is needed.

Argument source is either a state_initializer to create fresh replicas, or FromCheckpoint to load from a saved checkpoint.

Pigeons.create_state_initializer — Method

create_state_initializer(target, inputs)

Return a state_initializer used to populate the states at the beginning of the first round of Parallel Tempering.

Pigeons.create_swap_graph — Method

create_swap_graph(swap_graphs, shared)

Given a swap_graphs and Shared, return the swap_graph for the current iteration.

Pigeons.create_tempering — Method

create_tempering(inputs)

Build the tempering need for communicate!().

Pigeons.create_vector_replicas — Method

create_vector_replicas(inputs, shared, source)

Create replicas when distributed computing is not needed. See also state_initializer.

See create_replicas.

Pigeons.deo — Method

deo(potential, initial_state, initial_index, initial_lift, schedule, ϕ, 
    nscan, N, resolution, optimreference_round, modref_means, modref_stds, modref_covs, 
    full_covariance, prior_sampler, n_explore)

Deterministic even-odd parallel tempering (DEO/NRPT).

Arguments

potential: Function as in NRPT, but with only two arguments: x and η

initial_state: Starting state, as in NRPT. Input is of size: N+1 [ dim_x ]

initial_index: Starting indices

initial_lift: Starting lift

schedule: Annealing schedule

ϕ: As in NRPT

nscan: Number of scans to use

N: As in NRPT

resolution: As in NRPT

optimreference_round: As in NRPT

modref_means

modref_stds

prior_sample

Pigeons.deo — Method

deo(n_chains)

Implements the Deterministic Even Odd (DEO) scheme proposed in Okabe, 2001 and analyzed in Syed et al., 2021.

Pigeons.deoscan — Method

deoscan(potential, state, index, lift, etas, n, N, kernels, 
    optimreference_round, modref_means, modref_stds, modref_covs, full_covariance, 
    prior_sampler, n_explore)

Perform one DEO scan (local exploration + communication). Arguments are similar to those for deo(). Note that state is the state from the one previous scan, which is of size N+1[dim_x].

Pigeons.deserialize_immutables — Method

deserialize_immutables(filename)

See Immutable().

Pigeons.discretize — Method

discretize(path, betas)

Create log_potentials from a path by interpolating the path at each grid point specified in the Schedule.

Pigeons.entangler — Method

entangler(replicas)

Return the replicas's Entangler (possibly a no-communication Entangler if a single process is involved)

Pigeons.equally_spaced_schedule — Method

equally_spaced_schedule(n_chains)

Create a Schedule with n_chains equally spaced grid points.

Pigeons.explore! — Method

explore!(pt, explorer, multithreaded)

The @thread macro brings a large overhead even when Threads.nthreads == 1, so treating each case separately

Pigeons.explore! — Method

explore!(pt, explorer, multithreaded_flag)

Call sample_iid! or step!() on each chain (depending if it is a reference or not respectively).

Uses @threads to parallelize across threads. This is safe by the contract described in sample_iid! and step!().

Pigeons.explorer_recorder_builders — Method

explorer_recorder_builders(explorer)

What information is needed to perform adapt_explorer? Answer this by specifying an iterator containing recorder_builder's. Return [] if none are needed.

Pigeons.find_global_index — Method

find_global_index(lb, local_idx)

Find the global index corresponding to the given local_index.

Pigeons.find_local_index — Method

find_local_index(lb, global_idx)

Find the local index corresponding to the given global_index. Assumes the given global_index is one of this process'.

Pigeons.find_log_potential — Method

find_log_potential(replica, shared)

Find the log_potential for the chain the replica is at, based on the Shared object.

Pigeons.find_process — Method

find_process(lb, global_idx)

Find the process id (1-indexed) responsible for the given global_idx.

Pigeons.flush_immutables! — Method

flush_immutables!()

See Immutable().

Pigeons.index_process — Method

Full index process stored in memory.

Pigeons.index_process_plot — Method

index_process_plot(recorders)

Given a recorders, create an index process plot.

Pigeons.informal_doc — Method

informal_doc(doc_dir, mod)

Generate informal interface documentation, e.g.:

makedocs(;
    ...
    pages=[
        "Home" => "index.md", 
        "Interfaces" => informal_doc(@__DIR__, MyModuleName),
        ...
    ]
)

Pigeons.initialization — Method

initialization(state_initializer, rng, replica_index)

Determine state_initializer's initialization for the given replica_index.

Pigeons.interpolate — Method

interpolate(path, beta)

Returns the log_potential at point beta in the path.

Pigeons.is_finished — Method

is_finished(checkpoint_folder, inputs)

Is the provided path to a checkpoint folder complete? I.e. check in the .signal subfolder that all MPI processes have signaled that they are done.

Pigeons.latest_checkpoint_folder — Method

latest_checkpoint_folder(exec_folder)

Pigeons.load — Method

load(replicas)

Return the replicas's LoadBalance (possibly single_process_load)

Pigeons.load — Method

load(replicas)
load(result)

Load the result in memory.

Pigeons.local_exploration — Method

local_exploration(states, kernels, optimreference_round, modref_means, modref_stds, 
    modref_covs, full_covariance, prior_sampler, n_explore)
    chainacceptance = Vector{Int64}(undef, length(states))

Perform one local exploration move. state is the state from the one previous scan, which is of size N+1[dim_x].

Pigeons.locals — Method

locals(replicas)

Return the replica's that are stored in this machine

Pigeons.log_unnormalized_ratio — Method

log_unnormalized_ratio(
    log_potentials::AbstractVector,
    numerator::Int64,
    denominator::Int64,
    state
) -> Any

Assumes the input log_potentials is a vector where each element is a log_potential.

This default implementation is sufficient in most cases, but in less standard scenarios, e.g. where the state space is infinite dimensional, this can be overridden.

Pigeons.log_unnormalized_ratio — Method

log_unnormalized_ratio(
    log_potentials,
    numerator,
    denominator,
    state
)

The argument numerator selects one distribution $\pi_i$ from the collection log_potentials, and similarly denominator selects $\pi_j$. Let $x$ denote the input state. The ratio:

\[f(x) = \frac{\text{d}\pi_i}{\text{d}\pi_j}(x)\]

may only be known up to a normalization constant which can depend on $i$ and $j$ but not $x$, $g(x) = C_{i,j} f(x)$.

This function should return $\log g$ evaluated at state.

Pigeons.lognormalizingconstant — Method

lognormalizingconstant(energies, schedule)

Compute an estimate of the log normalizing constant given a vector of energies and the corresponding annealing schedule.

Pigeons.mpi_active — Method

mpi_active()

Detect if more than one MPI processes can be found.

Pigeons.my_global_indices — Method

my_global_indices(lb)

The slice of lb.global_indices this process is reponsible for.

Pigeons.my_load — Method

my_load(lb::Pigeons.LoadBalance) -> Int64

Return the number of indices (task) this process is responsible for.

Pigeons.n_chains — Method

n_chains(log_potentials)

The number of chains in the log_potentials.

Pigeons.next_exec_folder — Method

Return a unique subfolder of results/all/, making sure the unique folder and its parents are created. It will also create a soft symlink to it called results/latest`

Pigeons.one_per_host — Method

For benchmarking purpose: subset the communicator so that at most one MPI process runs in each machine.

Division is done so that original rank 0 is always included.

Return the new communicator or nothing if this machine is not in the subset.

See also '-s' option in mpi-run

Pigeons.only_one_process — Method

only_one_process(task, pt)

A task that should be ran on only one of the processes. Using the do .. end syntax, this can be used as:

only_one_process(pt) do 
    ...
end

Pigeons.partner_chain — Method

partner_chain(swap_graph, chain)

For a given swap_graph and input chain index, what chain will it interact with at the current iteration? Convention: if a chain is not interacting, return its index.

Pigeons.permuted_get — Method

permuted_get(p, indices)

Retreive the values for the given indices, using MPI communication when needed.

We make the following assumptions:

length(indices) == my_load(p.entangler.load)

the indices across all participating processes form a permutation of the global indices.

Pigeons.permuted_set! — Method

permuted_set!(p, indices, new_values)

Set the values for the given indices to the given new_values, using MPI communication when needed.

We make the same assumptions as in permuted_get().

Pigeons.pigeons — Method

pigeons(pt_arguments, new_process)

Run Parallel Tempering in a new process. See ChildProcess.

Pigeons.pigeons — Method

pigeons(pt_arguments, mpi_submission)

Pigeons.pigeons — Method

pigeons(pt_arguments; on)

pt_arguments can be either an Inputs, to start a new Parallel Tempering algorithm, or a string pointing to an execution to resume.

Pigeons.pigeons — Method

pigeons(pt)

Run (a generalization of) Parallel Tempering.

This will call several rounds of run_one_round!(), performing adaptation between each round via adapt().

This will also call report(), write_checkpoint(), and run_checks() between rounds.

Pigeons.pigeons — Method

pigeons(; on, args...)

Passes the args... to Inputs and start a new Parallel Tempering algorithm with that inputs.

Pigeons.providers — Method

providers(mod, name)

Provides a Set{Expr} containing all the providers of the given name in the given module.

Pigeons.record! — Method

record!(recorder, value)

Add value to the statistics accumulated by recorder.

Pigeons.record! — Method

record!(recorder, value)

Forwards to OnlineStats' fit!.

Pigeons.record! — Method

record!(recorder, value)

Given a value, a pair (a, b), and a Dict{K, Vector{V}} backed recorder, append b to the vector corresponding to a, inserting an empty vector into the dictionary first if needed.

Pigeons.record_if_requested! — Method

record_if_requested!(recorders, recorder_key, value)

If the recorders contains the given recorder_key, send the value to the recorder corresponding to the recorder_key. Otherwise, do nothing.

Pigeons.record_swap_stats! — Method

record_swap_stats!(
    pair_swapper,
    recorders,
    chain1,
    stat1,
    chain2,
    stat2
)

Given a pair_swapper, a recorders, the provided chain indices, and the sufficient statistics computed by swap_stat(), record statistics.

To avoid accumulating twice the same statistic with (chain1, chain2) and (chain2, chain2), swap!() only calls this for the pair with chain1 < chain2.

By default, the following are accumulated:

the swap acceptance probability.

TODO: stepping stone statistics.

Pigeons.record_swap_stats! — Method

record_swap_stats!(
    swapper,
    recorder,
    chain1,
    stat1,
    chain2,
    stat2
)

See TestSwapper.

Pigeons.reduce_deterministically — Method

reduce_deterministically(operation, source_data, e)

Perform a binary reduction of the source_data, using MPI when needed.

Consider the binary tree with leaves given by the global indices specified in e.load and stored in the different MPI processes' input source_data vectors. At each node of the tree, a reduction is performed using operation, i.e. by calling operation(left_child, right_child). When, and only when a branch of the tree crosses from one MPI process to another one, MPI communication is used to transmit the intermediate reduction.

At the end, for process 1, reduce_deterministically() will return the root of the binary tree, and for the other processes, reduce_deterministically() will return nothing.

Note that even when the operation is only approximately associative (typical situation for floating point reductions), the output of this function is invariant to the number of MPI processes involved (hence the terminology 'deterministically'). This contrasts to direct use of MPI collective communications where the leaves are MPI processes and hence will give slightly different outputs given different numbers of MPI processes. In the context of randomized algorithms, these minor differences are then amplified.

In contrast to transmit!(), we do not assume isbitstype(T) == true and use serialization when messages are transmitted over MPI.

Pigeons.reduce_recorders! — Method

reduce_recorders!(replicas)

Perform a reduction across all the replicas' individual recorders, using merge() on each individual recorder held. Returns a recorders with all the information merged.

Will reset the replicas' recorders at the same time using empty!().

Since this uses all_reduce_deterministically, the output is identical, no matter how many MPI processes are used, even when the reduction involves only approximately associative merge() operations (e.g. most floating point ones).

Pigeons.reference_chains — Method

reference_chains(swap_graphs, shared)

Given a swap_graphs and Shared, return a Set{Int} of chain(s) indices targetting the distribution of interest.

Pigeons.report — Method

report(pt, reduced_recorders)

Report summary information on the progress of pigeons().

Pigeons.restarts — Method

restarts(indices_matrix; cumulative) -> Any

Compute the number of restarts for a given index process trajectory. Otherwise, it is the same as roundtrip().

Pigeons.roundtrip — Method

roundtrip(indices_matrix; cumulative) -> Any

Compute the number of round trips for a given index process trajectory. indices_matrix is a matrix containing information about the index process. cumulative indicates whether we should store the output as a vector containing information about the number of total round trips up to sample n. If false, the output is a scalar.

Pigeons.run_checks — Method

Perform checks to detect software defects. Unable via field checked_round in Inputs Currently the following checks are implemented:

check_against_serial()

Pigeons.run_one_round! — Method

run_one_round!(pt)

From a PT object, run one round of a generalized version of Algorithm 1 in Syed et al., 2021.

Alternates between communicate!(), which consists of any pairwise communicating moves and [explore!()], which consists in moves independ to each chain.

Concrete specification of how to communicate and explore are specified by the field of type Shared contained in the provided PT.

Pigeons.sample_iid! — Method

sample_iid!(reference_log_potential, replica)

Perform i.i.d. sampling on the given Replica during its visit to the referencelogpotential created by create_reference_log_potential().

Pigeons.serialize_immutables — Method

serialize_immutables(filename)

See Immutable().

Pigeons.setkernels — Method

setkernels(potential, etas)

Set the local exploration kernels given the potential and the annealing parameters, etas.

Pigeons.setup_mpi — Method

setup_mpi(settings)

Run this function once before running MPI jobs. The setting are permanently saved

Pigeons.setup_mpi — Method

setup_mpi(; args...)

Arguments are passed in the constructor of MPISettings.

Pigeons.single_process_load — Method

single_process_load(n_global_indices)

A load balance with only one process.

Pigeons.slice_accept — Method

slice_accept(
    h,
    state,
    new_position,
    z,
    L,
    R,
    pointer,
    log_potential
)

Test whether to accept the current slice.

Pigeons.slice_double — Method

slice_double(h, state, z, pointer, log_potential, rng)

Double the current slice.

Pigeons.slice_sample! — Method

slice_sample!(h, state, log_potential, rng)

Slice sample one point.

Pigeons.slice_shrink — Method

slice_shrink(h, state, z, L, R, pointer, log_potential, rng)

Shrink the current slice.

Pigeons.sort_includes! — Method

sort_includes!(main)

Heuristic to automate the process of sorting include()'s.

Topological sorting of the source files under src (excluding main) is attempted, if successful, print the include string to copy and paste to the main file, otherwise, print the detected loops.

Internally, this function will:

Construct a graph where the vertices are the .jl files under src, excluding the provided main file (i.e. where the module is defined and the includes will sit in).

Each file starting with a capital letter is assumed to contain a struct with the same name as the file after removal of the .jl suffix. Similarly, files starting with @ are assumed to contain a macro with the similarly obtained name.

Each source file is inspected to see if the above struct and macro strings are detected. This defines edges in the graph. (known limitation: this includes spurious edges when e.g. the string occurs in a comment).

Pigeons.split_slice — Method

split_slice(slice, rng)

From one splittable random object, one can conceptualize an infinite list of splittable random objects. Return a slice from this infinite list.

Pigeons.step! — Method

step!(explorer, replica, shared)

Perform a transition on the given Replica invariant with respect to the distribution of the replica's chain.

The input explorer and Shared should only be read, not written to.

See also find_log_potential.

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

For each pair of chains encoded in swap_graph, use pair_swapper to decide if the pair will swap or not, and write the changes in-place into replicas (i.e. exchanging the Replica's chain fields for those that swapped.)

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

Entangled MPI swap! implementation.

This implementation is designed to support distributed PT with the following guarantees

The running time is independent of the size of the state space ('swapping annealing parameters rather than states')

The output is identical no matter how many MPI processes are used. In particular, this means that we can check correctness by comparing to the serial, single-process version.

Scalability to 1000s of processes communicating over MPI (see details below).

The same function can be used when a single process is used and MPI is not available.

Flexibility to extend PT to e.g. networks of targets and general paths.

Running time analysis:

Let $N$ denote the number of chains, $P$, the number of processes, and $K = \text{ceil}(N/P)$, the maximum number of chains held by one process. Assuming the running time is dominated by communication latency and a constant time for the latency of each peer-to-peer communication, the theoretical running time is $O(K)$. In practice, latency will grow as a function of $P$, but empirically, this growth appears to be slow enough that for say $P = N =$ a few 1000s, swapping will not be the computational bottleneck.

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

Single process, non-allocating swap! implementation.

Pigeons.swap_acceptance_pr — Method

Average MH swap acceptance probabilities for each pairs of interacting chains.

Pigeons.swap_decision — Method

swap_decision(pair_swapper, chain1, stat1, chain2, stat2)

Given a pair_swapper, a recorders, the provided chain indices, and the sufficient statistics computed by swap_stat(), make a swap decision.

By default, this is done as follows:

compute the standard swap acceptance probability min(1, exp(stat1.log_ratio + stat2.log_ratio))

make sure the two chains share the same uniform by picking the uniform from the chain with the smallest chain index

swap if the shared uniform is smaller than the swap acceptance probability.

Pigeons.swap_decision — Method

swap_decision(swapper, chain1, stat1, chain2, stat2)

See TestSwapper.

Pigeons.swap_stat — Method

swap_stat(pair_swapper, replica, partner_chain)

By default, two sufficient statistics are computed and stored in the SwapStat struct:

The result of calling log_unnormalized_ratio() on pair_swapper

A uniform number to coordinate the swap decision.

This can be extended by dispatching on other pair_swapper types, with the constraint that the returned sufficient statistics should satisfy isbitstype().

Pigeons.swap_stat — Method

swap_stat(swapper, replica, partner_chain)

See TestSwapper.

Pigeons.target_chains — Method

target_chains(swap_graphs, shared)

Given a swap_graphs and Shared, return the set of chain(s) targetting the reference distribution. These are typically tractable in the sense that we can sample i.i.d. from them.

Pigeons.tempering_recorder_builders — Method

tempering_recorder_builders(tempering)

What information is needed to perform adapt_tempering? Answer this by specifying an iterator containing recorder_builder's. Return [] if none are needed.

Pigeons.toy_mvn_target — Method

toy_mvn_target(
    dim::Int64
) -> Pigeons.ScaledPrecisionNormalPath

A toy multi-variate normal (mvn) target distribution used for testing. Uses a specialized path, ScaledPrecisionNormalPath, such that i.i.d. sampling is possible at all chains (via ToyExplorer).

Pigeons.transmit! — Method

transmit!(
    e,
    source_data,
    to_global_indices,
    write_received_data_here
)

Use MPI point-to-point communication to permute the contents of source_data across MPI processes, writing the permuted data into write_received_data_here. The permutation is specified by the load balance in the input argument e as well as the argument to_global_indices.

More precisely, assume the Vectors source_data, to_global_indices, and write_received_data_here are all of the length specified in my_load(e.load).

For each i, source_data[i] is sent to MPI process p = find_process(e.load, g), where g = to_global_indices[i] and written into this p 's write_received_data_here[j], where j = find_local_index(e.load, g)

See Entangler's comments regarding the requirement that all machines call transmit() the same number of times and at logically related intervals.

Additionally, at each micro-iteration, we assume that {to_global_indices_p : p ranges over the different processes} forms a partition of {1, ..., e.load.n_global_indices} If ran in single-process mode, this 'partition property' is checked; if ran in multi-process, opportunistic checks will be made, namely when several entries in to_global_indices lie in the same process, but systematic checks are not made for performance reasons.

We also assume isbitstype(T) == true.

Pigeons.transmit — Method

transmit(e, source_data, to_global_indices)

The same as transmit!() but instead of writing the result to an input argument, provide the result as a returned Vector.

Pigeons.updateschedule — Method

updateschedule(cumulativebarrier, N)

Update the annealing schedule. Given the cumulative communication barrier function in cumulativebarrier, find the optimal schedule of size N+1.

Pigeons.winsorized_mean — Method

winsorized_mean(x; α)

Compute the winsorized mean from an input x, which is assumed to be a vector of vectors. α denotes the percentage of observations to winsorize at the bottom and the top so that we use 1 - 2α observations and winsorize the rest.

Pigeons.winsorized_std — Method

winsorized_std(x; α)

Compute the winsorized standard deviation. The parameters are the same as those for winsorized_mean().

Pigeons.write_checkpoint — Method

write_checkpoint(pt, reduced_recorders)

If pt.inputs.checkpoint == true, save a checkpoint under [pt.exec_folder]/[unique folder]/round=[x]/checkpoint.

By default, pt.exec_folder is results/all/[unique folder].

In an MPI context, each MPI process will write its local replicas, while only one of the MPI processes will write the Shared and reduced recorders data. Moreover, only one MPI process will write once at the first round the Inputs data.

In cases where the sampled model contains large immutable data, consider using Immutable() to save disk space (Immutables will be written only by one MPI process at the first round).

Pigeons.@abstract — Macro

my_fct() = @abstract()

Define an abstract function (i.e. which gives an error message if calling it is attempted).

Pigeons.@informal — Macro

@informal name begin ... end

Document an informal interface with provided name, and functions specified in a begin .. end block.

@informal will spit back the contents of the begin .. end block so this macro can be essentially ignored at first read.

When building documentation, this allows us to use the function informal_doc() to automatically document the informal interface.

Pigeons.@weighted — Macro

@weighted(w, x)

Compute w*x, but if w==0.0, do not evaluate x and just return w (i.e. zero). Useful when x is computationally costly.