Index

Pigeons.explorer

Pigeons.log_potential

Pigeons.log_potentials

Pigeons.pair_swapper

Pigeons.path

Pigeons.recorder

Pigeons.recorder_builder

Pigeons.recorders

Pigeons.replicas

Pigeons.state

Pigeons.state_initializer

Pigeons.swap_graph

Pigeons.swap_graphs

Pigeons.target

Pigeons.tempering

Pigeons.var_reference

Pigeons.BlangTarget

Pigeons.ChildProcess

Pigeons.EntangledReplicas

Pigeons.Entangler

Pigeons.FromCheckpoint

Pigeons.GaussianReference

Pigeons.Immutable

Pigeons.Indexer

Pigeons.Indexer

Pigeons.Inputs

Pigeons.InterpolatedLogPotential

Pigeons.InterpolatingPath

Pigeons.Iterators

Pigeons.LoadBalance

Pigeons.MPI

Pigeons.MPISettings

Pigeons.NonReversiblePT

Pigeons.NonReversiblePT

Pigeons.OnlineStateRecorder

Pigeons.PT

Pigeons.PT

Pigeons.PT

Pigeons.PermutedDistributedArray

Pigeons.Replica

Pigeons.Result

Pigeons.RoundTripRecorder

Pigeons.ScaledPrecisionNormalPath

Pigeons.ScaledPrecisionNormalPath

Pigeons.Schedule

Pigeons.Shared

Pigeons.Shared

Pigeons.SliceSampler

Pigeons.StreamState

Pigeons.StreamTarget

Pigeons.Submission

Pigeons.SwapStat

Pigeons.TestSwapper

Pigeons.ThisProcess

Pigeons.ToyExplorer

Pigeons.TuringLogPotential

Pigeons.VarReference

Pigeons.activate_var_reference

Pigeons.adapt

Pigeons.adapt_explorer

Pigeons.adapt_tempering

Pigeons.adapted_schedule

Pigeons.all_reduce_deterministically

Pigeons.allocation_extrema

Pigeons.analytic_cumulativebarrier

Pigeons.blang_ising

Pigeons.blang_ising

Pigeons.blang_sitka

Pigeons.blang_sitka

Pigeons.check_against_serial

Pigeons.checksum

Pigeons.communicate!

Pigeons.communication_barriers

Pigeons.communicator

Pigeons.continuous_variables

Pigeons.continuous_variables

Pigeons.create_entangled_replicas

Pigeons.create_explorer

Pigeons.create_explorer

Pigeons.create_explorer

Pigeons.create_pair_swapper

Pigeons.create_path

Pigeons.create_path

Pigeons.create_recorders

Pigeons.create_recorders

Pigeons.create_reference_log_potential

Pigeons.create_replicas

Pigeons.create_state_initializer

Pigeons.create_swap_graph

Pigeons.create_tempering

Pigeons.create_var_reference

Pigeons.create_vector_replicas

Pigeons.default_recorder_builders

Pigeons.deo

Pigeons.deserialize_immutables

Pigeons.discrete_variables

Pigeons.discretize

Pigeons.energy_ac1

Pigeons.energy_ac1s

Pigeons.energy_ac1s

Pigeons.entangler

Pigeons.equally_spaced_schedule

Pigeons.explore!

Pigeons.explore!

Pigeons.explorer_recorder_builders

Pigeons.find_global_index

Pigeons.find_local_index

Pigeons.find_log_potential

Pigeons.find_process

Pigeons.flush_immutables!

Pigeons.index_process

Pigeons.informal_doc

Pigeons.initialization

Pigeons.initialization

Pigeons.interpolate

Pigeons.is_finished

Pigeons.is_reference

Pigeons.is_target

Pigeons.kill_job

Pigeons.last_round_max_allocation

Pigeons.last_round_max_time

Pigeons.latest_checkpoint_folder

Pigeons.load

Pigeons.load

Pigeons.locals

Pigeons.log_sum_ratio

Pigeons.log_unnormalized_ratio

Pigeons.log_unnormalized_ratio

Pigeons.mpi_active

Pigeons.my_global_indices

Pigeons.my_load

Pigeons.n_chains

Pigeons.n_round_trips

Pigeons.n_round_trips

Pigeons.n_tempered_restarts

Pigeons.n_tempered_restarts

Pigeons.next_exec_folder

Pigeons.number_of_chains

Pigeons.one_per_host

Pigeons.online_recorder_builders

Pigeons.only_one_process

Pigeons.optimal_schedule

Pigeons.partner_chain

Pigeons.permuted_get

Pigeons.permuted_set!

Pigeons.pigeons

Pigeons.pigeons

Pigeons.pigeons

Pigeons.pigeons

Pigeons.pigeons

Pigeons.providers

Pigeons.queue_status

Pigeons.queue_status

Pigeons.record!

Pigeons.record!

Pigeons.record!

Pigeons.record_if_requested!

Pigeons.record_swap_stats!

Pigeons.record_swap_stats!

Pigeons.reduce_deterministically

Pigeons.reduce_recorders!

Pigeons.report

Pigeons.round_trip

Pigeons.run_checks

Pigeons.run_one_round!

Pigeons.safelink

Pigeons.sample_iid!

Pigeons.sample_iid!

Pigeons.serialize_immutables

Pigeons.setup_blang

Pigeons.setup_mpi

Pigeons.setup_mpi

Pigeons.single_process_load

Pigeons.slice_accept

Pigeons.slice_double

Pigeons.slice_sample!

Pigeons.slice_shrink

Pigeons.sort_includes!

Pigeons.split_slice

Pigeons.step!

Pigeons.stepping_stone_pair

Pigeons.swap!

Pigeons.swap!

Pigeons.swap!

Pigeons.swap_acceptance_pr

Pigeons.swap_decision

Pigeons.swap_decision

Pigeons.swap_stat

Pigeons.swap_stat

Pigeons.target_online

Pigeons.tempering_recorder_builders

Pigeons.timing_extrema

Pigeons.toy_mvn_target

Pigeons.transmit

Pigeons.transmit!

Pigeons.update_reference!

Pigeons.update_state!

Pigeons.var_reference_recorder_builders

Pigeons.variable

Pigeons.watch

Pigeons.winsorized_mean

Pigeons.winsorized_std

Pigeons.write_checkpoint

RecipesBase.apply_recipe

Statistics.mean

Statistics.var

Pigeons.@abstract

Pigeons.@informal

Types and functions

Pigeons.BlangTarget — Type

A StreamTarget delegating exploration to Blang worker processes.

using Pigeons

Pigeons.setup_blang("blangDemos", "UBC-Stat-ML") # pre-compile the blang models in the github repo UBC-Stat-ML/blangDemos
pigeons(target = Pigeons.blang_ising());

Type Pigeons.blang followed by tab to find other examples.

Pigeons.ChildProcess — Type

Flag to run to a new julia process. Useful e.g. to dynamically control the number of threads to use. Fields:

n_threads: The number of threads to provide in the child julia process, the same as the current process by default.

dependencies: Julia modules (if of type Module) or paths to include (if of type String) needed by the child process.

n_local_mpi_processes: If greater than one, run the code locally over MPI using that many MPI processes. In most cases, this is useful only for debugging purpose, as multi-threading should typically perform better. This could also potentially be useful if using a third-party target distribution which somehow does not support multi-threading.

wait: If wait is false, the process runs asynchronously. When wait is false, the process' I/O streams are directed to devnull.

mpiexec_args: Extra arguments passed to mpiexec.

Pigeons.EntangledReplicas — Type

An implementation of replicas for distributed PT. Contains:

locals: The subset of replicas hosted in this process

chain_to_replica_global_indices: A specialized distributed array that maps chain indices to replica indices (global indices). This corresponds to the mapping $\boldsymbol{j}$ in line 2 of Algorithm 5 in Syed et al, 2021.

Pigeons.Entangler — Type

Assume all the MPI processes linked by this communicator will all call the key operations listed below the same number of times in their lifetime, at logically related occasions (e.g. a set number of times per iteration for algorithms running the same number of iterations). We call these 'occasions' a micro-iteration.

This datastructure keeps track internally of appropriate unique tags to coordinate the communication between MPI processes without having to do any explicit synchronization.

This struct contains:

communicator: An MPI Comm object (or nothing if a single process is involved).

load: How a set of tasks or "global indices" are distributed across processes.

current_received_bits: An internal datastructure used during MPI calls.

n_transmits: The current micro-iteration. Do not rely on it to count logical steps as it is reset to zero after transmit_counter_bound micor-iterations to avoid underflows to negative tags which cause MPI to crash.

transmit_counter_bound: Calculated from MPI.tagub and nglobal_indices to ensure MPI tags stay valid (i.e. do not overflow into negative values).

The key operations supported:

transmit() and transmit!(): encapsulates pairwise communications in which each MPI process is holding a Vector, the elements of which are to be permuted across the processes.

all_reduce_deterministically and reduce_deterministically, to perform MPI collective reduction while maintaining the Parallelism Invariance property.

Pigeons.FromCheckpoint — Type

Flag create_replicas (and related functions) that replicas should be loaded from a checkpoint. Fields:

checkpoint_folder

Pigeons.GaussianReference — Type

A Gaussian mean-field variational reference (i.e., with a diagonal covariance matrix).

Pigeons.Immutable — Method

Immutable(data)

Consider a situation where a distributed system serializes its state, and part of the state contains large immutable data. When the distributed processes each independently call Serialization.serialize(), naively the processes would each write identical copies of the large immutable data, which is space-inefficient.

Immutable resolves this space-inefficiency. For most users, all they need to know is to enclose large data inside the struct Immutable.

Details of how serialization/deserialization is performed:

Enclose large immutable data inside a Immutable. Assume the type of data has well defined hash and ==. Internally, we maintain an internal, global Dict indexed by hash(data) storing the data. This global Dict is called immutables.

Use flush_immutable() to clear the global immutable state

Use Serialization.serialize as usual. Internally, we dispatch serialization of Immutable is modified to skip the field containing the data.

Make one of the processes call serialize_immutables(). This serializes the immutables Dict.

Then for de-serialization, each process should call deserialize_immutables(). This restores immutable.

Finally, call Serialization.deserialize() as usual. When an Immutable instances is being deserialize, we dispatch deserialization so that the data is retreived from immutable.

Pigeons.Indexer — Type

A bijection between integers and some type T. T is assumed to have consistent hash and ==. The two sides of the bijection can be obtained with the fields:

i2t: A Vector mapping integers to objects t of type T.

t2i: A Dict mapping objects t of type T to integers.

Pigeons.Indexer — Method

Indexer(i2t)

Create an Indexer with the given Int to T mapping.

Pigeons.Inputs — Type

A Base.@kwdef struct used to create Parallel Tempering algorithms.

Fields (see source file for default values):

target: The target distribution.

seed: The master random seed.

n_rounds: The number of rounds to run.

n_chains: The number of chains to use.

n_chains_var_reference: The number of chains to use for the variational reference leg.

var_reference: The variational reference family.

checkpoint: Whether a checkpoint should be written to disk at the end of each round.

recorder_builders: An Vector with elements of type recorder_builder.

checked_round: The round index where run_checks() will be performed. Set to 0 to skip these checks.

multithreaded: If multithreaded explorers should be allowed. False by default since it incurs an overhead.

Pigeons.InterpolatedLogPotential — Type

A log_potential obtained by evaluation of a path at a point beta in the closed interval $[0, 1]$.

Pigeons.InterpolatingPath — Method

InterpolatingPath(ref, target)

Given a reference log_potential and a target log_potential, return a path interpolating between them.

By default, the interpolator is a LinearInterpolator, i.e. standard annealing.

Pigeons.Iterators — Type

Iterators used in Parallel Tempering. Stored in a struct so that recorder's can access it when outputting sample statistics.

Fields:

round: Index of the Parallel Tempering adaptation round, as defined in Algorithm 4 of Syed et al., 2021. Set to zero when when pigeons() not yet started.

scan: Number of (exploration, communication) pairs performed so far, corresponds to $n$ in Algorithm 1 of Syed et al., 2021. Round $i$ typically performs $2^i$ scans. Set to zero when runoneround!() is not yet started.

Pigeons.LoadBalance — Type

Split a list of indices across processes. These indices are denoted $1, 2, .., N$. They are usually some kind of task, for example in the context of parallel tempering, two kinds of tasks arise:

in replicas.state, task $i$ consists in keeping track of the state of replica $i$.

in replicas.chain_to_replica_global_indices, task $i$ consists in storing which replica index corresponds to chain $i$.

One such task index is called a global_index.

LoadBalance splits the global indices among n_processes. LoadBalance is constructed so that the difference in the number of global indices a process is responsible of (its "load") is at most one.

A LoadBalance contains:

my_process_index: A unique index for this process. We use 1-indexed, i.e. hide MPI's 0-indexed ranks.

n_processes: Total number of processes involved.

n_global_indices: The total number of global indices shared between all the processes.

The set {1, 2, .., load()} is called a set of local indices. A local index indexes a slice in {1, 2, ..., n_global_indices}. Collectively over the n_processes, these slices form a partition of the global indices.

Key functions to utilize a LoadBalance struct:

my_global_indices()

find_process()

find_local_index()

my_load()

Pigeons.MPI — Type

Flag to run on MPI. Before using, you have to call once setup_mpi.

Fields:

n_threads: The number of threads per MPI process, 1 by default.

walltime: The walltime limit, 00:30:00 by default (i.e., 30 minutes).

n_mpi_processes: The number of MPI processes, 2 by default.

memory: The memory allocated to each MPI process, 8gb by default.

dependencies: Julia modules (if of type Module) or paths to include (if of type String) needed by the child process.

mpiexec_args: Extra arguments passed to mpiexec.

Pigeons.MPISettings — Type

Global settings needed for MPI job submission:

allocation_code: E.g., for -A in PBS submission scripts.

environment_modules: "Envirnonment modules" to load (not to be confused with Julia modules). Run module avail in the HPC login node to see what is available on your HPC. For example: ["git", "gcc", "intel-mkl", "openmpi"]

Pigeons.NonReversiblePT — Type

Variables needed for the non-reversible Parallel Tempering described in Syed et al., 2021:

path: The path.

schedule: The Schedule.

log_potentials: The log_potentials.

swap_graphs: The swap_graphs.

communication_barriers: The communication barriers computed by communication_barriers() at the same time as this tempering was created; or nothing before adaptation, i.e. before the first call to adapt_tempering.

Pigeons.NonReversiblePT — Method

NonReversiblePT(inputs)

The adaptive non-reversible Parallel Tempering described in Syed et al., 2021.

Pigeons.OnlineStateRecorder — Type

See target_online().

Pigeons.PT — Type

Storage involved in PT algorithms:

inputs: The user-provided Inputs that determine the execution of a PT algorithm.

replicas: The replicas held by this machine.

shared: Information shared across all machines, updated between rounds.

exec_folder: Either a path to a folder shared by all processes, which is used to save information to disk (checkpoints, samples etc); or nothing if a completely in-memory algorithm is used.

reduced_recorders: recorders from the last round, or empty recorders.

Pigeons.PT — Method

PT(source_exec_folder; round, fresh_exec_folder)

Create a PT struct from a saved checkpoint. The path [checkpoint_folder] should point to a folder with the name checkpoint found under results/all/[exec_folder]/round=x.

The checkpoint carries all the information stored in a PT struct. It is possible for an MPI-based execution to load a checkpoint written by a single-process execution and vice versa.

A new unique folder will be created with symlinks to the source one, so that e.g. running more rounds of PT will results in a new space-efficient checkpoint containing all the information for the new run.

Pigeons.PT — Method

PT(inputs; exec_folder)

Create a PT struct from provided Inputs. Optionally, provide a specific exec_folder path (AbstractString), if not one will be created via next_exec_folder().

Pigeons.PermutedDistributedArray — Type

A distributed array making special assumptions on how it will be accessed and written to. The indices of this distributed array correspond to the notion of "global indices" defined in LoadBalance. Several MPI processes cooperate, each processing storing data for a slice of this distributed array.

We make the following assumptions:

Each MPI process will set/get entries the same number of times in their lifetime, at logically related episodes (e.g. a set number of times per iteration for algorithms running the same number of iterations). These episodes are called micro-iterations as in Entangler, which this datastructure is built on.

Moreover, at each time all processes perform a get or a set, we assume that each global index is manipulated by exactly one process (i.e. an implicit permutation of the global indices).

We use these assumptions to achieve read/write costs that are near-constant in the number of machines participating.

This struct contains:

local_data: The slice of the distributed array maintained by this MPI process.

entangler: An Entangler used to coordinate communication.

The operations supported are:

permuted_get()

permuted_set!()

Pigeons.Replica — Type

One of the $N$ components that forms the state maintained by a PT algorithm. A Replica contains:

state: Configuration in the state space.

chain: The index of the distribution currently associated with this replica, modified during swaps.

rng: Random operations involving this state should use only this random number generator.

recorders: Records statistics. Each replica carries its own for thread safety/distribution; then they are reduced at end of each round.

replica_index: A global id associated with this replica.

Pigeons.Result — Type

A link to an execution folder able to deserialize type T via a string constructor.

Pigeons.RoundTripRecorder — Type

See round_trip().

Pigeons.ScaledPrecisionNormalPath — Type

A path of zero-mean normals for testing; contains:

precision0: Precision parameter of the reference.

precision1: Precision parameter of the target.

dim: Dimensionality.

Pigeons.ScaledPrecisionNormalPath — Method

ScaledPrecisionNormalPath(dim)

Toy Multivariate Normal (MVN) path of distributions for testing: see section I.4.1 in Syed et al 2021.

Pigeons.Schedule — Type

A partition of $[0, 1]$ encoded by monotonically increasing grid points starting at zero and ending at one.

Pigeons.Shared — Type

Information shared by all processes involved in a round of distributed parallel tempering. This is updated between rounds but only read during a round.

Fields:

iterators: See Iterators.

tempering: See tempering.

explorer: See explorer.

var_reference: See var_reference

Only one instance maintained per process.

Pigeons.Shared — Method

Shared(inputs)

Create a Shared struct based on an Inputs.

Uses create_tempering() and create_explorer().

Pigeons.SliceSampler — Type

Slice sampler based on Neal, 2003.

Pigeons.StreamState — Type

States used in the replicas when a StreamTarget is used.

Pigeons.StreamTarget — Type

A target based on running worker processes, one for each replica, each communicating with Pigeons using standard streams. These worker processes can be implemented in an arbitrary programming language.

StreamTarget implements log_potential and explorer by invoking worker processes via standard stream communication. The standard stream is less efficient than alternatives such as protobuff, but it has the advantage of being supported by nearly all programming languages in existence. Also in many practical cases, since the worker process is invoked only three times per chain per iteration, it is unlikely to be the bottleneck (overhead is in the order of 0.1ms).

The worker process should be able to reply to commands of the following forms (one command per line):

log_potential(0.6) in the worker's stdin to which it should return a response of the form response(-124.23) in its stdout, providing in this example the joint log density at beta = 0.6;

call_sampler!(0.4) signaling that one round of local exploration should be performed at beta = 0.4, after which the worker should signal it is done with response().

Pigeons.Submission — Type

Specifies where to submit a task.

Pigeons.SwapStat — Type

Default statistics exchanged by a pair of chains in the process of proposing a swap:

log_ratio

uniform

See swap_stat()

Pigeons.TestSwapper — Type

For testing/benchmarking purposes, a simple pair_swapper where all swaps have equal acceptance probability.

Could also be used to warm-start swap connections during exploration phase by setting that constant probability to zero.

Pigeons.ThisProcess — Type

Flag to ask to run a function within the current process.

Pigeons.ToyExplorer — Type

Toy explorer for toy paths where each log_potential supports i.i.d. sampling via rand!(rng, x, log_potential).

Pigeons.TuringLogPotential — Method

TuringLogPotential(model)

Given a DynamicPPL.Model from Turing.jl, create a TuringLogPotential conforming both target and log_potential.

Pigeons.VarReference — Type

Abstract type for variational references.

Pigeons.activate_var_reference — Method

activate_var_reference(var_reference, iterators)

Choose on which rounds/scans to activate the variational reference.

Pigeons.adapt — Method

adapt(pt, reduced_recorders)

Call adapt_tempering() followed by adapt_explorer.

Pigeons.adapt_explorer — Method

adapt_explorer(explorer, reduced_recorders, shared)

Called between successive rounds (run_one_round!).

Given an explorer, reduced recorders and Shared return an updated explorer.

Pigeons.adapt_tempering — Method

adapt_tempering(
    tempering,
    reduced_recorders,
    iterators,
    var_reference,
    state
)

Called between successive rounds (run_one_round!).

Given a tempering and reduced recorders return an updated tempering.

Pigeons.adapted_schedule — Method

adapted_schedule(n_chains, cumulativebarrier)

Create a Schedule with n_chains grid points computed using Algorithm 2 in Syed et al, 2021.

Pigeons.all_reduce_deterministically — Method

all_reduce_deterministically(operation, source_data, e)

Same as reduce_deterministically() except that the result at the root of the tree is then broadcast to all machines so that the output of all_reduce_deterministically() is the root of the reduction tree for all MPI processes involved.

Pigeons.allocation_extrema — Method

Allocations informations.

Pigeons.analytic_cumulativebarrier — Method

analytic_cumulativebarrier(path)

Known cumulative barrier used for testing, from Predescu et al., 2003.

Pigeons.blang_ising — Method

blang_ising(model_options)

Two-dimensional Ising model.

For more information:

using Pigeons

Pigeons.setup_blang("blangDemos") 
run(Pigeons.blang_ising(`--help`).command);

E.g., use arguments `model.N` to set the size 
of the grid. 

Pigeons.blang_ising — Method

blang_ising()

15x15 Ising model.

Pigeons.blang_sitka — Method

blang_sitka(model_options)

Model for phylogenetic inference from single-cell copy-number alteration from Salehi et al., 2020.

For more information:

using Pigeons

Pigeons.setup_blang("nowellpack") 
run(Pigeons.blang_sitka(`--help`).command);

Pigeons.blang_sitka — Method

blang_sitka()

Default options for infering a posterior distribution on phylogenetic trees for the 535 triple negative breast cancer dataset in Salehi et al., 2020.

Pigeons.check_against_serial — Method

check_against_serial(pt)

Run a separate, fully serial version of the PT algorithm, and compare the checkpoint files to ensure the two produce exactly the same output.

Pigeons.checksum — Function

checksum(filename)
checksum(filename, blocksize)

Given a filename path, compute a crc32c() checksum in constant memory.

Pigeons.communicate! — Method

communicate!(pt)

Use create_pair_swapper() and create_swap_graph to construct the inputs needed for swap!.

Pigeons.communication_barriers — Method

communication_barriers(intensity, schedule)

Compute the local communication barrier and cumulative barrier functions from the intensity rates (i.e. rejection rates in the context of Parallel Tempering) and the current annealing schedule. The estimation of the barriers is based on Fritsch-Carlson monotonic interpolation.

Returns a NamedTuple with fields:

localbarrier

cumulativebarrier

globalbarrier

Pigeons.communicator — Method

communicator(replicas)

Return the replicas's MPI.Comm or nothing if no MPI needed

Pigeons.continuous_variables — Method

continuous_variables(state)

The names (each a Symbol) of the continuous variables in the given state.

Pigeons.continuous_variables — Method

continuous_variables(pt)

Pigeons.create_entangled_replicas — Method

create_entangled_replicas(inputs, shared, source)

Create distributed replicas.

See create_replicas.

Pigeons.create_explorer — Method

create_explorer(target, inputs)

Current explorer for Turing models: a SliceSampler. Slice sampler has the advantage of being not very sensitive to tuning.

Pigeons.create_explorer — Method

create_explorer(target, inputs)

Create an explorer for the given target.

Pigeons.create_explorer — Method

create_explorer(inputs)

Given an Inputs object, dispatch on create_explorer(inputs.target, inputs) to construct the explorer associated with the input target distribution.

Pigeons.create_pair_swapper — Method

create_pair_swapper(tempering, target)

Given a tempering and a Shared struct, create a pair_swapper.

If omitted, by default will return the standard Metropolis-Hastings accept-reject.

Pigeons.create_path — Method

create_path(target, inputs)

Create a path, by default linking the given target to the refence provided by create_reference_log_potential().

For this default to work, the target should conform both target and log_potential.

Pigeons.create_path — Method

create_path(target, inputs)

In this case, the target is already a path, so return it.

Pigeons.create_recorders — Method

create_recorders(recorder_builders)

Create a recorders from an iterable with element type recorder_builder.

Pigeons.create_recorders — Method

create_recorders(inputs, shared)

Create a recorders from an Inputs and Shared.

Pigeons.create_reference_log_potential — Method

create_reference_log_potential(target, inputs)

Create a default reference distribution, by returning a log_potential. The returned object will also get passed to sample_iid!() at the "hot chains" of the Parallel Tempering algorithm.

Pigeons.create_replicas — Method

create_replicas(inputs, shared, source)

Create replicas, detecting automatically if MPI is needed.

Argument source is either a state_initializer to create fresh replicas, or FromCheckpoint to load from a saved checkpoint.

Pigeons.create_state_initializer — Method

create_state_initializer(target, inputs)

Return a state_initializer used to populate the states at the beginning of the first round of Parallel Tempering.

Pigeons.create_swap_graph — Method

create_swap_graph(swap_graphs, shared)

Given a swap_graphs and Shared, return the swap_graph for the current iteration.

Pigeons.create_tempering — Method

create_tempering(inputs)

Build the tempering needed for communicate!().

Pigeons.create_var_reference — Method

Create a variational reference.

Pigeons.create_vector_replicas — Method

create_vector_replicas(inputs, shared, source)

Create replicas when distributed computing is not needed. See also state_initializer.

See create_replicas.

Pigeons.default_recorder_builders — Method

Set of recorders with no measurable impact on performance.

Pigeons.deo — Method

deo(n_chains)

Implements the Deterministic Even Odd (DEO) scheme proposed in Okabe, 2001 and analyzed in Syed et al., 2021.

Pigeons.deserialize_immutables — Method

deserialize_immutables(filename)

See Immutable().

Pigeons.discrete_variables — Method

discrete_variables(state)

The names (each a Symbol) of the discrete (Int) variables in the given state.

Pigeons.discretize — Method

discretize(path, betas)

Create log_potentials from a path by interpolating the path at each grid point specified in the Schedule.

Pigeons.energy_ac1 — Method

Auto-correlation before and after an exploration step, grouped by chain.

Pigeons.energy_ac1s — Function

energy_ac1s(reduced_recorders)
energy_ac1s(reduced_recorders, skip_reference)
energy_ac1s(reduced_recorders, skip_reference, pt)

Pigeons.energy_ac1s — Function

energy_ac1s(pt)
energy_ac1s(pt, skip_reference)

Auto-correlations between energy before and after an exploration step, for each chain. Organized as a Vector where component i corresponds to chain i.

It is often useful to skip the reference chain, for two reasons, first, exploration should be iid there, second, if the prior is flat the auto-correlation of the energy will be NaN for the reference.

Pigeons.entangler — Method

entangler(replicas)

Return the replicas's Entangler (possibly a no-communication Entangler if a single process is involved)

Pigeons.equally_spaced_schedule — Method

equally_spaced_schedule(n_chains)

Create a Schedule with n_chains equally spaced grid points.

Pigeons.explore! — Method

explore!(pt, explorer, multithreaded)

The @threads macro brings a large overhead even when Threads.nthreads == 1 (!), so a separate method is used for the single thread mode.

Pigeons.explore! — Method

explore!(pt, explorer, multithreaded_flag)

Call sample_iid! or step!() on each chain (depending if it is a reference or not respectively).

Uses @threads to parallelize across threads. This is safe by the contract described in sample_iid!() and step!().

Pigeons.explorer_recorder_builders — Method

explorer_recorder_builders(explorer)

What information is needed to perform adapt_explorer? Answer this by specifying an iterator containing recorder_builder's. Return [] if none are needed.

Pigeons.find_global_index — Method

find_global_index(lb, local_idx)

Find the global index corresponding to the given local_index.

Pigeons.find_local_index — Method

find_local_index(lb, global_idx)

Find the local index corresponding to the given global_index. Assumes the given global_index is one of this process'.

Pigeons.find_log_potential — Method

find_log_potential(replica, shared)

Find the log_potential for the chain the replica is at, based on the Shared object.

Pigeons.find_process — Method

find_process(lb, global_idx)

Find the process id (1-indexed) responsible for the given global_idx.

Pigeons.flush_immutables! — Method

flush_immutables!()

See Immutable().

Pigeons.index_process — Method

Full index process stored in memory.

Pigeons.informal_doc — Method

informal_doc(doc_dir, mod)

Generate informal interface documentation, e.g.:

makedocs(;
    ...
    pages=[
        "Home" => "index.md", 
        "Interfaces" => informal_doc(@__DIR__, MyModuleName),
        ...
    ]
)

Pigeons.initialization — Method

initialization(state_initializer, rng, replica_index)

Determine state_initializer's initialization for the given replica_index.

Pigeons.initialization — Method

initialization(target, rng, replica_index)

Return StreamState by following these steps:

create a Cmd that uses the provided rng to set the random seed properly, as well as target-specific configurations provided by target.

Create StreamState from the Cmd created in step 1 and return it.

Pigeons.interpolate — Method

interpolate(path, beta)

Returns the log_potential at point beta in the path.

Pigeons.is_finished — Method

is_finished(checkpoint_folder, inputs)

Is the provided path to a checkpoint folder complete? I.e. check in the .signal subfolder that all MPI processes have signaled that they are done.

Pigeons.is_reference — Method

is_reference(swap_graphs, chain)

Pigeons.is_target — Method

is_target(swap_graphs, chain)

Pigeons.kill_job — Method

kill_job(result)

Instruct the scheduler to cancel or kill a job.

Pigeons.last_round_max_allocation — Method

Maximum bytes allocated (over the MPI process) to compute the last Parallel Tempering round.

Pigeons.last_round_max_time — Method

Maximum time (over the MPI process) to compute the last Parallel Tempering round.

Pigeons.latest_checkpoint_folder — Method

latest_checkpoint_folder(exec_folder)

Pigeons.load — Method

load(replicas)

Return the replicas's LoadBalance (possibly single_process_load)

Pigeons.load — Method

load(replicas)
load(result)

Load the result in memory.

Pigeons.locals — Method

locals(replicas)

Return the replica's that are stored in this machine

Pigeons.log_sum_ratio — Method

Log of the sum of density ratios between neighbour chains, used to compute stepping stone estimators of lognormalization contants.

Pigeons.log_unnormalized_ratio — Method

log_unnormalized_ratio(
    log_potentials,
    numerator,
    denominator,
    state
)

Assumes the input log_potentials is a vector where each element is a log_potential.

This default implementation is sufficient in most cases, but in less standard scenarios, e.g. where the state space is infinite dimensional, this can be overridden.

Pigeons.log_unnormalized_ratio — Method

log_unnormalized_ratio(
    log_potentials,
    numerator,
    denominator,
    state
)

The argument numerator selects one distribution $\pi_i$ from the collection log_potentials, and similarly denominator selects $\pi_j$. Let $x$ denote the input state. The ratio:

\[f(x) = \frac{\text{d}\pi_i}{\text{d}\pi_j}(x)\]

may only be known up to a normalization constant which can depend on $i$ and $j$ but not $x$, $g(x) = C_{i,j} f(x)$.

This function should return $\log g$ evaluated at state.

Pigeons.mpi_active — Method

mpi_active()

A flag is set by launch scripts (see ChildProcess.jl) to indicate if this process is a child MPI process under an mpiexec. Otherwise, that flag is false by default.

This function retrieves the value of that flag.

Pigeons.my_global_indices — Method

my_global_indices(lb)

The slice of lb.global_indices this process is reponsible for.

Pigeons.my_load — Method

my_load(lb)

Return the number of indices (task) this process is responsible for.

Pigeons.n_chains — Method

n_chains(log_potentials)

The number of chains in the log_potentials.

Pigeons.n_round_trips — Method

n_round_trips(reduced_recorders)

Pigeons.n_round_trips — Method

n_round_trips(pt)

Pigeons.n_tempered_restarts — Method

n_tempered_restarts(reduced_recorders)

Pigeons.n_tempered_restarts — Method

n_tempered_restarts(pt)

Pigeons.next_exec_folder — Method

Return a unique subfolder of results/all/, making sure the unique folder and its parents are created. It will also create a soft symlink to it called results/latest`

Pigeons.number_of_chains — Method

Extract the number of Parallel Tempering chains from Inputs.

Pigeons.one_per_host — Method

For benchmarking purpose: subset the communicator so that at most one MPI process runs in each machine.

Division is done so that original rank 0 is always included.

Return the new communicator or nothing if this machine is not in the subset.

See also '-s' option in mpi-run

Pigeons.online_recorder_builders — Method

Set of constant memory recorders.

Pigeons.only_one_process — Method

only_one_process(task, pt)

A task that should be ran on only one of the processes. Using the do .. end syntax, this can be used as:

only_one_process(pt) do 
    ...
end

Pigeons.optimal_schedule — Method

optimal_schedule(
    intensity,
    old_schedule,
    new_schedule_n_chains
)

Return an optimal Schedule based on statistics from a previous round.

Pigeons.partner_chain — Method

partner_chain(swap_graph, chain)

For a given swap_graph and input chain index, what chain will it interact with at the current iteration? Convention: if a chain is not interacting, return its index.

Pigeons.permuted_get — Method

permuted_get(p, indices)

Retreive the values for the given indices, using MPI communication when needed.

We make the following assumptions:

length(indices) == my_load(p.entangler.load)

the indices across all participating processes form a permutation of the global indices.

Pigeons.permuted_set! — Method

permuted_set!(p, indices, new_values)

Set the values for the given indices to the given new_values, using MPI communication when needed.

We make the same assumptions as in permuted_get().

Pigeons.pigeons — Method

pigeons(pt_arguments, new_process)

Run Parallel Tempering in a new process. See ChildProcess.

Pigeons.pigeons — Method

pigeons(pt_arguments, mpi_submission)

Pigeons.pigeons — Method

pigeons(pt_arguments; on)

pt_arguments can be either an Inputs, to start a new Parallel Tempering algorithm, or a string pointing to an execution to resume.

Pigeons.pigeons — Method

pigeons(pt)

Run (a generalization of) Parallel Tempering.

This will call several rounds of run_one_round!(), performing adaptation between each round via adapt().

This will also call report(), write_checkpoint(), and run_checks() between rounds.

Pigeons.pigeons — Method

pigeons(; on, args...)

Passes the args... to Inputs and start a new Parallel Tempering algorithm with that inputs.

Pigeons.providers — Method

providers(mod, name)

Provides a Set{Expr} containing all the providers of the given name in the given module.

Pigeons.queue_status — Method

queue_status(result)

Display the queue status for one MPI job.

Pigeons.queue_status — Method

queue_status()

Display the queue status for all the user's jobs.

Pigeons.record! — Method

record!(recorder, value)

Add value to the statistics accumulated by recorder.

Pigeons.record! — Method

record!(recorder, value)

Forwards to OnlineStats' fit!.

Pigeons.record! — Method

record!(recorder, value)

Given a value, a pair (a, b), and a Dict{K, Vector{V}} backed recorder, append b to the vector corresponding to a, inserting an empty vector into the dictionary first if needed.

Pigeons.record_if_requested! — Method

record_if_requested!(recorders, recorder_key, value)

If the recorders contains the given recorder_key, send the value to the recorder corresponding to the recorder_key. Otherwise, do nothing.

Pigeons.record_swap_stats! — Method

record_swap_stats!(
    pair_swapper,
    recorders,
    chain1,
    stat1,
    chain2,
    stat2
)

Given a pair_swapper, a recorders, the provided chain indices, and the sufficient statistics computed by swap_stat(), record statistics.

To avoid accumulating twice the same statistic with (chain1, chain2) and (chain2, chain2), swap!() only calls this for the pair with chain1 < chain2.

Pigeons.record_swap_stats! — Method

record_swap_stats!(
    swapper,
    recorder,
    chain1,
    stat1,
    chain2,
    stat2
)

See TestSwapper.

Pigeons.reduce_deterministically — Method

reduce_deterministically(operation, source_data, e)

Perform a binary reduction of the source_data, using MPI when needed.

Consider the binary tree with leaves given by the global indices specified in e.load and stored in the different MPI processes' input source_data vectors. At each node of the tree, a reduction is performed using operation, i.e. by calling operation(left_child, right_child). When, and only when a branch of the tree crosses from one MPI process to another one, MPI communication is used to transmit the intermediate reduction.

At the end, for process 1, reduce_deterministically() will return the root of the binary tree, and for the other processes, reduce_deterministically() will return nothing.

Note that even when the operation is only approximately associative (typical situation for floating point reductions), the output of this function is invariant to the number of MPI processes involved (hence the terminology 'deterministically'). This contrasts to direct use of MPI collective communications where the leaves are MPI processes and hence will give slightly different outputs given different numbers of MPI processes. In the context of randomized algorithms, these minor differences are then amplified.

In contrast to transmit!(), we do not assume isbitstype(T) == true and use serialization when messages are transmitted over MPI.

Pigeons.reduce_recorders! — Method

reduce_recorders!(replicas)

Perform a reduction across all the replicas' individual recorders, using Base.merge() on each individual recorder held. Returns a recorders with all the information merged.

Will reset the replicas' recorders at the same time using Base.empty!().

Since this uses all_reduce_deterministically, the output is identical, no matter how many MPI processes are used, even when the reduction involves only approximately associative Base.merge() operations (e.g. most floating point ones).

Pigeons.report — Method

report(pt)

Report summary information on the progress of pigeons().

Pigeons.round_trip — Method

Restart and round-trip counts.

Pigeons.run_checks — Method

Perform checks to detect software defects. Unable via field checked_round in Inputs Currently the following checks are implemented:

check_against_serial()

Pigeons.run_one_round! — Method

run_one_round!(pt)

From a PT object, run one round of a generalized version of Algorithm 1 in Syed et al., 2021.

Alternates between communicate!(), which consists of any pairwise communicating moves and [explore!()], which consists in moves independ to each chain.

Concrete specification of how to communicate and explore are specified by the field of type Shared contained in the provided PT.

Pigeons.safelink — Method

safelink(target, link)

Work around two issues with symlink():

naively calling symlink() when there are relative paths leads to broken links

on windows, one needs admin permission to do symlinks, so print a helpful error message in that case

Pigeons.sample_iid! — Method

sample_iid!(reference_log_potential, replica)

Perform i.i.d. sampling on the given Replica during its visit to the referencelogpotential created by create_reference_log_potential().

Pigeons.sample_iid! — Method

sample_iid!(var_reference, replica)

Obtain one iid sample from the reference distribution specified by the variational family.

Pigeons.serialize_immutables — Method

serialize_immutables(filename)

See Immutable().

Pigeons.setup_blang — Function

setup_blang(repo_name)
setup_blang(repo_name, organization)

Download the github repo with the given repo_name and organization in ~.pigeons, and compile the blang code.

Pigeons.setup_mpi — Method

setup_mpi(settings)

Run this function once before running MPI jobs. The setting are permanently saved. See MPISettings.

Pigeons.setup_mpi — Method

setup_mpi(; args...)

Arguments are passed in the constructor of MPISettings.

Pigeons.single_process_load — Method

single_process_load(n_global_indices)

A load balance with only one process.

Pigeons.slice_accept — Method

slice_accept(
    h,
    state,
    new_position,
    z,
    L,
    R,
    pointer,
    log_potential
)

Test whether to accept the current slice.

Pigeons.slice_double — Method

slice_double(h, state, z, pointer, log_potential, rng)

Double the current slice.

Pigeons.slice_sample! — Method

slice_sample!(h, state, log_potential, rng)

Slice sample one point.

Pigeons.slice_shrink — Method

slice_shrink(h, state, z, L, R, pointer, log_potential, rng)

Shrink the current slice.

Pigeons.sort_includes! — Method

sort_includes!(main)

Heuristic to automate the process of sorting include()'s.

Topological sorting of the source files under src (excluding main) is attempted, if successful, print the include string to copy and paste to the main file, otherwise, print the detected loops.

Internally, this function will:

Construct a graph where the vertices are the .jl files under src, excluding the provided main file (i.e. where the module is defined and the includes will sit in).

Each file starting with a capital letter is assumed to contain a struct with the same name as the file after removal of the .jl suffix. Similarly, files starting with @ are assumed to contain a macro with the similarly obtained name.

Each source file is inspected to see if the above struct and macro strings are detected. This defines edges in the graph. (known limitation: this includes spurious edges when e.g. the string occurs in a comment).

Pigeons.split_slice — Method

split_slice(slice, rng)

From one splittable random object, one can conceptualize an infinite list of splittable random objects. Return a slice from this infinite list.

Pigeons.step! — Method

step!(explorer, replica, shared)

Perform a transition on the given Replica invariant with respect to the distribution of the replica's chain.

The input explorer and Shared should only be read, not written to.

See also find_log_potential.

Pigeons.stepping_stone_pair — Method

stepping_stone_pair(pt)

Assuming that the reference distribution has a normalization constant of one, compute the (log of) the stepping stone estimator. It returns a pair, one such that its exponential is unbiased under Assumptions (A1-2) in Syed et al., 2021 for $Z$ and the other, for $1/Z$. Both are consistent in the number of MCMC iterations without these strong assumptions.

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

For each pair of chains encoded in swap_graph, use pair_swapper to decide if the pair will swap or not, and write the changes in-place into replicas (i.e. exchanging the Replica's chain fields for those that swapped.)

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

Entangled MPI swap! implementation.

This implementation is designed to support distributed PT with the following guarantees

The running time is independent of the size of the state space ('swapping annealing parameters rather than states')

The output is identical no matter how many MPI processes are used. In particular, this means that we can check correctness by comparing to the serial, single-process version.

Scalability to 1000s of processes communicating over MPI (see details below).

The same function can be used when a single process is used and MPI is not available.

Flexibility to extend PT to e.g. networks of targets and general paths.

Running time analysis:

Let $N$ denote the number of chains, $P$, the number of processes, and $K = \text{ceil}(N/P)$, the maximum number of chains held by one process. Assuming the running time is dominated by communication latency and a constant time for the latency of each peer-to-peer communication, the theoretical running time is $O(K)$. In practice, latency will grow as a function of $P$, but empirically, this growth appears to be slow enough that for say $P = N =$ a few 1000s, swapping will not be the computational bottleneck.

Pigeons.swap! — Method

swap!(pair_swapper, replicas, swap_graph)

Single process, non-allocating swap! implementation.

Pigeons.swap_acceptance_pr — Method

Average MH swap acceptance probabilities for each pairs of interacting chains.

Pigeons.swap_decision — Method

swap_decision(pair_swapper, chain1, stat1, chain2, stat2)

Given a pair_swapper, a recorders, the provided chain indices, and the sufficient statistics computed by swap_stat(), make a swap decision.

By default, this is done as follows:

compute the standard swap acceptance probability min(1, exp(stat1.log_ratio + stat2.log_ratio))

make sure the two chains share the same uniform by picking the uniform from the chain with the smallest chain index

swap if the shared uniform is smaller than the swap acceptance probability.

Pigeons.swap_decision — Method

swap_decision(swapper, chain1, stat1, chain2, stat2)

See TestSwapper.

Pigeons.swap_stat — Method

swap_stat(pair_swapper, replica, partner_chain)

By default, two sufficient statistics are computed and stored in the SwapStat struct:

The result of calling log_unnormalized_ratio() on pair_swapper

A uniform number to coordinate the swap decision.

This can be extended by dispatching on other pair_swapper types, with the constraint that the returned sufficient statistics should satisfy isbitstype().

Pigeons.swap_stat — Method

swap_stat(swapper, replica, partner_chain)

See TestSwapper.

Pigeons.target_online — Method

Online statistics on the target chain.

Pigeons.tempering_recorder_builders — Method

tempering_recorder_builders(tempering)

What information is needed to perform adapt_tempering? Answer this by specifying an iterator containing recorder_builder's. Return [] if none are needed.

Pigeons.timing_extrema — Method

Timing informations.

Pigeons.toy_mvn_target — Method

toy_mvn_target(dim)

A toy multi-variate normal (mvn) target distribution used for testing. Uses a specialized path, ScaledPrecisionNormalPath, such that i.i.d. sampling is possible at all chains (via ToyExplorer).

Pigeons.transmit! — Method

transmit!(
    e,
    source_data,
    to_global_indices,
    write_received_data_here
)

Use MPI point-to-point communication to permute the contents of source_data across MPI processes, writing the permuted data into write_received_data_here. The permutation is specified by the load balance in the input argument e as well as the argument to_global_indices.

More precisely, assume the Vectors source_data, to_global_indices, and write_received_data_here are all of the length specified in my_load(e.load).

For each i, source_data[i] is sent to MPI process p = find_process(e.load, g), where g = to_global_indices[i] and written into this p 's write_received_data_here[j], where j = find_local_index(e.load, g)

See Entangler's comments regarding the requirement that all machines call transmit() the same number of times and at logically related intervals.

Additionally, at each micro-iteration, we assume that {to_global_indices_p : p ranges over the different processes} forms a partition of {1, ..., e.load.n_global_indices} If ran in single-process mode, this 'partition property' is checked; if ran in multi-process, opportunistic checks will be made, namely when several entries in to_global_indices lie in the same process, but systematic checks are not made for performance reasons.

We also assume isbitstype(T) == true.

Pigeons.transmit — Method

transmit(e, source_data, to_global_indices)

The same as transmit!() but instead of writing the result to an input argument, provide the result as a returned Vector.

Pigeons.update_reference! — Method

update_reference!(reduced_recorders, var_reference, state)

Update the variational reference and the annealing path. Returns the new annealing path.

Pigeons.update_state! — Method

update_state!(state, name, index, value)

Update the state's entry at symbol name and index with value.

Pigeons.var_reference_recorder_builders — Method

var_reference_recorder_builders(var_reference)

Specify the recorder builders for this variational reference family.

Pigeons.variable — Method

variable(state, name)

The storage within the state of the variable of the given name, typically an Array.

Pigeons.watch — Method

watch(result; machine, last, interactive)

Print the queue status as well as the standard out and error streams (merged) for the given machine.

Pigeons.winsorized_mean — Method

winsorized_mean(x; α)

Compute the winsorized mean from an input x, which is assumed to be a vector of vectors. α denotes the percentage of observations to winsorize at the bottom and the top so that we use 1 - 2α observations and winsorize the rest.

Pigeons.winsorized_std — Method

winsorized_std(x; α)

Compute the winsorized standard deviation. The parameters are the same as those for winsorized_mean().

Pigeons.write_checkpoint — Method

write_checkpoint(pt)

If pt.inputs.checkpoint == true, save a checkpoint under [pt.exec_folder]/[unique folder]/round=[x]/checkpoint.

By default, pt.exec_folder is results/all/[unique folder].

In an MPI context, each MPI process will write its local replicas, while only one of the MPI processes will write the Shared and reduced recorders data. Moreover, only one MPI process will write once at the first round the Inputs data.

In cases where the sampled model contains large immutable data, consider using Immutable() to save disk space (Immutables will be written only by one MPI process at the first round).

RecipesBase.apply_recipe — Method

apply_recipe(plotattributes, index_process)

A plot @recipe for an index process.

Statistics.mean — Function

mean(pt)
mean(pt, variable_name)

Statistics.var — Function

var(pt)
var(pt, variable_name)

Pigeons.@abstract — Macro

my_fct() = @abstract()

Define an abstract function (i.e. which gives an error message if calling it is attempted).

Pigeons.@informal — Macro

@informal name begin ... end

Document an informal interface with provided name, and functions specified in a begin .. end block.

@informal will spit back the contents of the begin .. end block so this macro can be essentially ignored at first read.

When building documentation, this allows us to use the function informal_doc() to automatically document the informal interface.